chore: remove unused code from warehouse
