+++ .github/workflows/tests.yaml
@@ -40,7 +40,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        destination: [ bigquery, clickhouse, datalake, deltalake, deltalake-native, mssql, azure-synapse, postgres, postgres-legacy, redshift, snowflake ]
+        destination: [ bigquery, clickhouse, datalake, deltalake, mssql, azure-synapse, postgres, redshift, snowflake ]
         include:
           - package: warehouse/integrations/bigquery
             destination: bigquery
@@ -48,22 +48,14 @@ jobs:
             destination: clickhouse
           - package: warehouse/integrations/datalake
             destination: datalake
-          - package: warehouse/integrations/deltalake-native
+          - package: warehouse/integrations/deltalake
             destination: deltalake
-            useNative: false
-          - package: warehouse/integrations/deltalake-native
-            destination: deltalake-native
-            useNative: true
           - package: warehouse/integrations/mssql
             destination: mssql
           - package: warehouse/integrations/azure-synapse
             destination: azure-synapse
           - package: warehouse/integrations/postgres
             destination: postgres
-            useLegacy: false
-          - package: warehouse/integrations/postgres
-            destination: postgres-legacy
-            useLegacy: true
           - package: warehouse/integrations/redshift
             destination: redshift
           - package: warehouse/integrations/snowflake
@@ -91,8 +83,6 @@ jobs:
           REDSHIFT_INTEGRATION_TEST_CREDENTIALS: ${{ secrets.REDSHIFT_INTEGRATION_TEST_CREDENTIALS }}
           SNOWFLAKE_INTEGRATION_TEST_CREDENTIALS: ${{ secrets.SNOWFLAKE_INTEGRATION_TEST_CREDENTIALS }}
           SNOWFLAKE_RBAC_INTEGRATION_TEST_CREDENTIALS: ${{ secrets.SNOWFLAKE_RBAC_INTEGRATION_TEST_CREDENTIALS }}
-          RSERVER_WAREHOUSE_POSTGRES_USE_LEGACY: ${{ matrix.useLegacy }}
-          RSERVER_WAREHOUSE_DELTALAKE_USE_NATIVE: ${{ matrix.useNative }}
       - name: Upload coverage report
         uses: actions/upload-artifact@v3
         with:
@@ -149,7 +139,7 @@ jobs:
   coverage:
     name: Coverage
     runs-on: 'ubuntu-20.04'
-    needs: 
+    needs:
      - warehouse-integration
      - unit
      - package-unit
