+++ utils/misc/misc.go
@@ -1,7 +1,6 @@
 package misc
 
 import (
-	"archive/zip"
 	"bufio"
 	"bytes"
 	"compress/gzip"
@@ -21,7 +20,6 @@ import (
 	"regexp"
 	"runtime"
 	"runtime/debug"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -38,8 +36,6 @@ import (
 
 	"github.com/rudderlabs/rudder-go-kit/config"
 	"github.com/rudderlabs/rudder-go-kit/logger"
-	"github.com/rudderlabs/rudder-go-kit/stats/metric"
-
 	"github.com/rudderlabs/rudder-server/utils/httputil"
 	"github.com/rudderlabs/rudder-server/utils/types"
 )
@@ -88,17 +84,6 @@ type RudderError struct {
 	StackTrace        string
 	Code              int
 }
-
-type pair struct {
-	key   string
-	value float64
-}
-
-type pairList []pair
-
-func (p pairList) Len() int           { return len(p) }
-func (p pairList) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
-func (p pairList) Less(i, j int) bool { return p[i].value < p[j].value }
 
 type RFP struct {
 	path         string
@@ -156,15 +141,6 @@ func saveErrorStore(errorStore ErrorStoreT) {
 	}
 }
 
-// AppendError creates or appends second error to first error
-func AppendError(callingMethodName string, firstError, secondError *error) {
-	if *firstError != nil {
-		*firstError = fmt.Errorf("%v ; %v : %w", (*firstError).Error(), callingMethodName, *secondError)
-	} else {
-		*firstError = fmt.Errorf("%v : %w", callingMethodName, *secondError)
-	}
-}
-
 // RecordAppError appends the error occurred to error_store.json
 func RecordAppError(err error) {
 	if err == nil {
@@ -219,77 +195,6 @@ func GetRudderEventVal(key string, rudderEvent types.SingularEventT) (interface{
 		return nil, false
 	}
 	return rudderVal, true
-}
-
-// GetRudderID return the UserID from the object
-func GetRudderID(event types.SingularEventT) (string, bool) {
-	userID, ok := GetRudderEventVal("rudderId", event)
-	if !ok {
-		// TODO: Remove this in next build.
-		// This is for backwards compatibility, esp for those with sessions.
-		userID, ok = GetRudderEventVal("anonymousId", event)
-		if !ok {
-			return "", false
-		}
-	}
-	userIDStr, ok := userID.(string)
-	return userIDStr, ok
-}
-
-// ZipFiles compresses files[] into zip at filename
-func ZipFiles(filename string, files []string) error {
-	newZipFile, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = newZipFile.Close() }()
-
-	zipWriter := zip.NewWriter(newZipFile)
-	defer func() { _ = zipWriter.Close() }()
-
-	// Add files to zip
-	for _, file := range files {
-		if err = AddFileToZip(zipWriter, file); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// AddFileToZip adds file to zip including size header stats
-func AddFileToZip(zipWriter *zip.Writer, filename string) error {
-	fileToZip, err := os.Open(filename)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = fileToZip.Close() }()
-
-	// Get the file information
-	info, err := fileToZip.Stat()
-	if err != nil {
-		return err
-	}
-
-	header, err := zip.FileInfoHeader(info)
-	if err != nil {
-		return err
-	}
-
-	// Using FileInfoHeader() above only uses the basename of the file. If we want
-	// to preserve the folder structure we can overwrite this with the full path.
-	// uncomment this line to preserve folder structure
-	// header.Name = filename
-
-	// Change to deflate to gain better compression
-	// see http://golang.org/pkg/archive/zip/#pkg-constants
-	header.Method = zip.Deflate
-
-	writer, err := zipWriter.CreateHeader(header)
-	if err != nil {
-		return err
-	}
-	_, err = io.Copy(writer, fileToZip)
-	return err
 }
 
 // RemoveFilePaths removes filePaths as well as cleans up the empty folder structure.
@@ -362,23 +267,6 @@ func RemoveEmptyFolderStructureForFilePath(fp string) {
 		}
 		currDir = parentDir
 	}
-}
-
-// ReadLines reads a whole file into memory
-// and returns a slice of its lines.
-func ReadLines(path string) ([]string, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = file.Close() }()
-
-	var lines []string
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	return lines, scanner.Err()
 }
 
 var logOnce sync.Once
@@ -460,14 +348,6 @@ func GetChronologicalTimeStamp(receivedAt, sentAt, originalTimestamp time.Time) 
 	return receivedAt.Add(-sentAt.Sub(originalTimestamp))
 }
 
-func MapStringKeys(input map[string]interface{}) []string {
-	keys := make([]string, 0, len(input))
-	for k := range input {
-		keys = append(keys, k)
-	}
-	return keys
-}
-
 func TruncateStr(str string, limit int) string {
 	if len(str) > limit {
 		str = str[:limit]
@@ -540,10 +420,6 @@ func MakeHTTPRequestWithTimeout(url string, payload io.Reader, timeout time.Dura
 	}
 
 	return respBody, resp.StatusCode, nil
-}
-
-func HTTPCallWithRetry(url string, payload []byte) ([]byte, int) {
-	return HTTPCallWithRetryWithTimeout(url, payload, time.Second*150)
 }
 
 func ConvertInterfaceToStringArray(input []interface{}) []string {
@@ -752,25 +628,6 @@ func GetMacAddress() string {
 	return macAddress.String()
 }
 
-func KeepProcessAlive() {
-	var ch chan int
-	<-ch
-}
-
-// GetOutboundIP returns preferred outbound ip of this machine
-// https://stackoverflow.com/a/37382208
-func GetOutboundIP() (net.IP, error) {
-	conn, err := net.Dial("udp", "8.8.8.8:80")
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = conn.Close() }()
-
-	localAddr := conn.LocalAddr().(*net.UDPAddr)
-
-	return localAddr.IP, nil
-}
-
 /*
 RunWithTimeout runs provided function f until provided timeout d.
 If the timeout is reached, onTimeout callback will be called.
@@ -899,22 +756,6 @@ func GetObjectStorageConfig(opts ObjectStorageOptsT) map[string]interface{} {
 	return objectStorageConfigMap
 }
 
-func GetSpacesLocation(location string) (region string) {
-	r, _ := regexp.Compile(`\.*.*\.digitaloceanspaces\.com`) // skipcq: GO-S1009
-	subLocation := r.FindString(location)
-	regionTokens := strings.Split(subLocation, ".")
-	if len(regionTokens) == 3 {
-		region = regionTokens[0]
-	}
-	return region
-}
-
-// GetNodeID returns the nodeId of the current node
-func GetNodeID() string {
-	nodeID := config.MustGetString("INSTANCE_ID")
-	return nodeID
-}
-
 // GetMD5UUID hashes the given string into md5 and returns it as auuid
 func GetMD5UUID(str string) (uuid.UUID, error) {
 	// To maintain backward compatibility, we are using md5 hash of the string
@@ -996,20 +837,6 @@ func GetMandatoryJSONFieldNames(st interface{}) []string {
 	return mandatoryJSONFieldNames
 }
 
-func MinInt(a, b int) int {
-	if a <= b {
-		return a
-	}
-	return b
-}
-
-func MaxInt(a, b int) int {
-	if a >= b {
-		return a
-	}
-	return b
-}
-
 // GetTagName gets the tag name using a uuid and name
 func GetTagName(id string, names ...string) string {
 	var truncatedNames string
@@ -1053,37 +880,6 @@ func GetWarehouseURL() (url string) {
 		url = fmt.Sprintf(`http://localhost:%d`, config.GetInt("Warehouse.webPort", 8082))
 	} else {
 		url = config.GetString("WAREHOUSE_URL", "http://localhost:8082")
-	}
-	return
-}
-
-func GetDatabricksVersion() (version string) {
-	url := fmt.Sprintf(`%s/databricksVersion`, GetWarehouseURL())
-	req, err := http.NewRequest(http.MethodGet, url, http.NoBody)
-	if err != nil {
-		return
-	}
-	client := &http.Client{
-		Timeout: config.GetDuration("HttpClient.timeout", 30, time.Second),
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		pkgLogger.Errorf("Unable to make a warehouse databricks build version call with error : %s", err.Error())
-		return
-	}
-	if resp == nil {
-		version = "No response from warehouse."
-		return
-	}
-	defer func() { httputil.CloseResponse(resp) }()
-	if resp.StatusCode == http.StatusOK {
-		bodyBytes, err := io.ReadAll(resp.Body)
-		if err != nil {
-			pkgLogger.Errorf("Unable to read response into bytes with error : %s", err.Error())
-			version = "Unable to read response from warehouse."
-			return
-		}
-		version = string(bodyBytes)
 	}
 	return
 }
@@ -1207,25 +1003,6 @@ func GetJsonSchemaDTFromGoDT(goType string) string {
 		return "boolean"
 	}
 	return "object"
-}
-
-func SortMap(inputMap map[string]metric.MovingAverage) []string {
-	pairArr := make(pairList, len(inputMap))
-
-	i := 0
-	for k, v := range inputMap {
-		pairArr[i] = pair{k, v.Value()}
-		i++
-	}
-
-	sort.Sort(pairArr)
-	var sortedWorkspaceList []string
-	// p is sorted
-	for _, k := range pairArr {
-		// Workspace ID - RS Check
-		sortedWorkspaceList = append(sortedWorkspaceList, k.key)
-	}
-	return sortedWorkspaceList
 }
 
 // SleepCtx sleeps for the given duration or until the context is canceled.
