+++ warehouse/constraint_test.go
@@ -25,11 +25,11 @@ var _ = Describe("Constraint", func() {
 		Entry(nil, warehouseutils.POSTGRES, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
 		Entry(nil, warehouseutils.CLICKHOUSE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
 		Entry(nil, warehouseutils.MSSQL, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
-		Entry(nil, warehouseutils.AZURE_SYNAPSE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
+		Entry(nil, warehouseutils.AzureSynapse, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
 		Entry(nil, warehouseutils.DELTALAKE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
-		Entry(nil, warehouseutils.S3_DATALAKE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
-		Entry(nil, warehouseutils.GCS_DATALAKE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
-		Entry(nil, warehouseutils.AZURE_DATALAKE, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
+		Entry(nil, warehouseutils.S3Datalake, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
+		Entry(nil, warehouseutils.GCSDatalake, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
+		Entry(nil, warehouseutils.AzureDatalake, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
 		Entry(nil, warehouseutils.BQ, &BatchRouterEvent{}, "id", &ConstraintsViolation{}),
 		Entry(nil, warehouseutils.BQ,
 			&BatchRouterEvent{
