+++ warehouse/integrations/deltalake/deltalake.go
@@ -2,52 +2,68 @@ package deltalake
 
 import (
 	"context"
+	"database/sql"
+	"errors"
 	"fmt"
 	"regexp"
+	"strconv"
 	"strings"
 	"time"
 
 	"golang.org/x/exp/slices"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
+
+	warehouseclient "github.com/rudderlabs/rudder-server/warehouse/client"
+	sqlmiddleware "github.com/rudderlabs/rudder-server/warehouse/integrations/middleware/sqlquerywrapper"
+
+	dbsqllog "github.com/databricks/databricks-sql-go/logger"
+
+	dbsql "github.com/databricks/databricks-sql-go"
+
+	"github.com/rudderlabs/rudder-server/warehouse/logfield"
 
 	"github.com/rudderlabs/rudder-go-kit/config"
 	"github.com/rudderlabs/rudder-go-kit/logger"
 	"github.com/rudderlabs/rudder-go-kit/stats"
-	proto "github.com/rudderlabs/rudder-server/proto/databricks"
 	"github.com/rudderlabs/rudder-server/utils/misc"
-	warehouseclient "github.com/rudderlabs/rudder-server/warehouse/client"
-	"github.com/rudderlabs/rudder-server/warehouse/integrations/deltalake/client"
 	"github.com/rudderlabs/rudder-server/warehouse/internal/model"
 	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
 
-// Database configuration
 const (
-	Host                   = "host"
-	Port                   = "port"
-	Path                   = "path"
-	Token                  = "token"
-	Catalog                = "catalog"
-	UseSTSTokens           = "useSTSTokens"
-	EnableExternalLocation = "enableExternalLocation"
-	ExternalLocation       = "externalLocation"
+	host         = "host"
+	port         = "port"
+	path         = "path"
+	token        = "token"
+	catalog      = "catalog"
+	useSTSTokens = "useSTSTokens"
+	userAgent    = "Rudderstack"
 )
 
 const (
-	provider       = warehouseutils.DELTALAKE
+	provider = warehouseutils.DELTALAKE
+
+	// Corresponds to the max length set for event rudder-transformer
+	// https://github.com/rudderlabs/rudder-transformer/blob/fb8b818b2cbd05f784117b9f3040856dab1a7346/src/warehouse/v1/util.js#L34
 	tableNameLimit = 127
 )
 
-// Reference: https://docs.oracle.com/cd/E17952_01/connector-odbc-en/connector-odbc-reference-errorcodes.html
 const (
-	tableOrViewNotFound = "42S02"
-	columnNotFound      = "42000"
-	databaseNotFound    = "42000"
-	partitionNotFound   = "42000"
+	schemaNotFound       = "[SCHEMA_NOT_FOUND]"
+	partitionNotFound    = "SHOW PARTITIONS is not allowed on a table that is not partitioned"
+	columnsAlreadyExists = "already exists in root"
 )
 
-// Rudder data type mapping with Delta lake mappings.
+const (
+	mergeMode  = "MERGE"
+	appendMode = "APPEND"
+)
+
+const (
+	rudderStagingTableRegex    = "^rudder_staging_.*$"       // matches rudder_staging_* tables
+	nonRudderStagingTableRegex = "^(?!rudder_staging_.*$).*" // matches tables that do not start with rudder_staging_
+)
+
+// dataTypesMap maps rudder data types to delta lake data types
 var dataTypesMap = map[string]string{
 	"boolean":  "BOOLEAN",
 	"int":      "BIGINT",
@@ -57,7 +73,7 @@ var dataTypesMap = map[string]string{
 	"date":     "DATE",
 }
 
-// Delta Lake mapping with rudder data types mappings.
+// dataTypesMapToRudder maps delta lake data types to rudder data types
 // Reference: https://docs.databricks.com/sql/language-manual/sql-ref-datatype-rules.html
 var dataTypesMapToRudder = map[string]string{
 	"TINYINT":   "int",
@@ -87,11 +103,10 @@ var dataTypesMapToRudder = map[string]string{
 // excludeColumnsMap Columns you need to exclude
 // Since event_date is an auto generated column in order to support partitioning.
 // We need to ignore it during query generation.
-var excludeColumnsMap = map[string]bool{
-	"event_date": true,
+var excludeColumnsMap = map[string]struct{}{
+	"event_date": {},
 }
 
-// Primary Key mappings for tables
 var primaryKeyMap = map[string]string{
 	warehouseutils.UsersTable:      "id",
 	warehouseutils.IdentifiesTable: "id",
@@ -110,7 +125,7 @@ var errorsMappings = []model.JobError{
 }
 
 type Deltalake struct {
-	Client                 *client.Client
+	DB                     *sqlmiddleware.DB
 	Namespace              string
 	ObjectStorage          string
 	Warehouse              model.Warehouse
@@ -118,660 +133,308 @@ type Deltalake struct {
 	ConnectTimeout         time.Duration
 	Logger                 logger.Logger
 	Stats                  stats.Stats
-	Schema                 string
-	SparkServerType        string
-	AuthMech               string
-	UID                    string
-	ThriftTransport        string
-	SSL                    string
-	UserAgent              string
-	GrpcTimeout            time.Duration
-	HealthTimeout          time.Duration
 	LoadTableStrategy      string
 	EnablePartitionPruning bool
-	ConnectorURL           string
+	SlowQueryThreshold     time.Duration
+	MaxRetries             int
+	RetryMinWait           time.Duration
+	RetryMaxWait           time.Duration
 }
 
 func New() *Deltalake {
 	return &Deltalake{
-		Logger: logger.NewLogger().Child("warehouse").Child("integrations").Child("deltalake"),
+		Logger: logger.NewLogger().Child("warehouse").Child("integration").Child("deltalake"),
 		Stats:  stats.Default,
 	}
 }
 
 func WithConfig(h *Deltalake, config *config.Config) {
-	h.Schema = config.GetString("Warehouse.deltalake.schema", "default")
-	h.SparkServerType = config.GetString("Warehouse.deltalake.sparkServerType", "3")
-	h.AuthMech = config.GetString("Warehouse.deltalake.authMech", "3")
-	h.UID = config.GetString("Warehouse.deltalake.uid", "token")
-	h.ThriftTransport = config.GetString("Warehouse.deltalake.thriftTransport", "2")
-	h.SSL = config.GetString("Warehouse.deltalake.ssl", "1")
-	h.UserAgent = config.GetString("Warehouse.deltalake.userAgent", "RudderStack")
-	h.GrpcTimeout = config.GetDuration("Warehouse.deltalake.grpcTimeout", 2, time.Minute)
-	h.HealthTimeout = config.GetDuration("Warehouse.deltalake.healthTimeout", 15, time.Second)
-	h.LoadTableStrategy = config.GetString("Warehouse.deltalake.loadTableStrategy", "MERGE")
+	h.LoadTableStrategy = config.GetString("Warehouse.deltalake.loadTableStrategy", mergeMode)
 	h.EnablePartitionPruning = config.GetBool("Warehouse.deltalake.enablePartitionPruning", true)
-	h.ConnectorURL = config.GetString("DATABRICKS_CONNECTOR_URL", "localhost:50051")
+	h.SlowQueryThreshold = config.GetDuration("Warehouse.deltalake.slowQueryThreshold", 5, time.Minute)
+	h.MaxRetries = config.GetInt("Warehouse.deltalake.maxRetries", 10)
+	h.RetryMinWait = config.GetDuration("Warehouse.deltalake.retryMinWait", 1, time.Second)
+	h.RetryMaxWait = config.GetDuration("Warehouse.deltalake.retryMaxWait", 300, time.Second)
 }
 
-// getDeltaLakeDataType returns datatype for delta lake which is mapped with rudder stack datatype
-func getDeltaLakeDataType(columnType string) string {
-	return dataTypesMap[columnType]
-}
-
-// ColumnsWithDataTypes returns columns with specified prefix and data type
-func ColumnsWithDataTypes(columns model.TableSchema, prefix string) string {
-	keys := warehouseutils.SortColumnKeysFromColumnMap(columns)
-	format := func(idx int, name string) string {
-		if _, ok := excludeColumnsMap[name]; ok {
-			return ""
-		}
-		if name == "received_at" {
-			generatedColumnSQL := "DATE GENERATED ALWAYS AS ( CAST(received_at AS DATE) )"
-			return fmt.Sprintf(`%s%s %s, %s%s %s`, prefix, name, getDeltaLakeDataType(columns[name]), prefix, "event_date", generatedColumnSQL)
-		}
-
-		return fmt.Sprintf(`%s%s %s`, prefix, name, getDeltaLakeDataType(columns[name]))
-	}
-	return warehouseutils.JoinWithFormatting(keys, format, ",")
-}
-
-// columnNames returns joined column with comma separated
-func columnNames(keys []string) string {
-	return strings.Join(keys, ",")
-}
-
-// stagingColumnNames returns staging column names
-func stagingColumnNames(keys []string) string {
-	format := func(idx int, str string) string {
-		return fmt.Sprintf(`STAGING.%s`, str)
-	}
-	return warehouseutils.JoinWithFormatting(keys, format, ",")
-}
-
-// columnsWithValues returns main and staging column values.
-func columnsWithValues(keys []string) string {
-	format := func(idx int, str string) string {
-		return fmt.Sprintf(`MAIN.%[1]s = STAGING.%[1]s`, str)
-	}
-	return warehouseutils.JoinWithFormatting(keys, format, ",")
-}
-
-// checkAndIgnoreAlreadyExistError checks and ignores native errors.
-func checkAndIgnoreAlreadyExistError(errorCode, ignoreError string) bool {
-	if errorCode == "" || errorCode == ignoreError {
-		return true
-	}
-	return false
-}
-
-// NewClient creates deltalake client
-func (dl *Deltalake) NewClient(ctx context.Context, cred *client.Credentials, connectTimeout time.Duration) (Client *client.Client, err error) {
-	identifier := misc.FastUUID().String()
-	connConfig := &proto.ConnectionConfig{
-		Host:            cred.Host,
-		Port:            cred.Port,
-		HttpPath:        cred.Path,
-		Pwd:             cred.Token,
-		Schema:          dl.Schema,
-		SparkServerType: dl.SparkServerType,
-		AuthMech:        dl.AuthMech,
-		Uid:             dl.UID,
-		ThriftTransport: dl.ThriftTransport,
-		Ssl:             dl.SSL,
-		UserAgentEntry:  dl.UserAgent,
-	}
-
-	// Getting timeout context
-	timeout := dl.GrpcTimeout
-	if connectTimeout != 0 {
-		timeout = connectTimeout
-	}
-	tCtx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	// Creating grpc connection using timeout context
-	conn, err := grpc.DialContext(tCtx, dl.ConnectorURL, grpc.WithTransportCredentials(insecure.NewCredentials()), grpc.WithBlock())
-	if err == context.DeadlineExceeded {
-		execTimeouts := dl.Stats.NewStat("warehouse.deltalake.grpcTimeouts", stats.CountType)
-		execTimeouts.Count(1)
-
-		err = fmt.Errorf("connection timed out to Delta lake: %w", err)
-		return
-	}
-	if err != nil {
-		err = fmt.Errorf("error while creating grpc connection to Delta lake: %w", err)
-		return
-	}
-
-	dbClient := proto.NewDatabricksClient(conn)
-	connectionResponse, err := dbClient.Connect(ctx, &proto.ConnectRequest{
-		Config:     connConfig,
-		Identifier: identifier,
-	})
-	if err != nil {
-		err = fmt.Errorf("error connecting to Delta lake: %w", err)
-		return
-	}
-	if connectionResponse.GetErrorCode() != "" {
-		err = fmt.Errorf("error connecting to Delta lake with response: %s", connectionResponse.GetErrorMessage())
-		return
-	}
-
-	Client = &client.Client{
-		Logger:         dl.Logger,
-		CredConfig:     connConfig,
-		CredIdentifier: identifier,
-		Conn:           conn,
-		Client:         dbClient,
-	}
-
-	// Setting up catalog at the client level
-	if catalog := warehouseutils.GetConfigValue(Catalog, dl.Warehouse); catalog != "" {
-		sqlStatement := fmt.Sprintf("USE CATALOG `%s`;", catalog)
-
-		if err = dl.ExecuteSQLClient(ctx, Client, sqlStatement); err != nil {
-			return
-		}
-	}
-	return
-}
-
-func (*Deltalake) DeleteBy(context.Context, []string, warehouseutils.DeleteByParams) error {
-	return fmt.Errorf(warehouseutils.NotImplementedErrorCode)
-}
-
-// fetchTables fetch tables with tableNames
-func (dl *Deltalake) fetchTables(ctx context.Context, dbT *client.Client, schema string) (tableNames []string, err error) {
-	fetchTableResponse, err := dbT.Client.FetchTables(ctx, &proto.FetchTablesRequest{
-		Config:     dbT.CredConfig,
-		Identifier: dbT.CredIdentifier,
-		Schema:     schema,
-	})
-	if err != nil {
-		return tableNames, fmt.Errorf("%s Error while fetching tables: %v", dl.GetLogIdentifier(), err)
-	}
-	if !checkAndIgnoreAlreadyExistError(fetchTableResponse.GetErrorCode(), databaseNotFound) {
-		err = fmt.Errorf("%s Error while fetching tables with response: %v", dl.GetLogIdentifier(), fetchTableResponse.GetErrorMessage())
-		return
-	}
-	tableNames = append(tableNames, fetchTableResponse.GetTables()...)
-	return
-}
-
-// fetchPartitionColumns return the partition columns for the corresponding tables
-func (dl *Deltalake) fetchPartitionColumns(ctx context.Context, dbT *client.Client, tableName string) ([]string, error) {
-	sqlStatement := fmt.Sprintf(`SHOW PARTITIONS %s.%s`, dl.Warehouse.Namespace, tableName)
-
-	columnsResponse, err := dbT.Client.FetchPartitionColumns(ctx, &proto.FetchPartitionColumnsRequest{
-		Config:       dbT.CredConfig,
-		Identifier:   dbT.CredIdentifier,
-		SqlStatement: sqlStatement,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch partition columns, error: %w", err)
-	}
-	if !checkAndIgnoreAlreadyExistError(columnsResponse.GetErrorCode(), partitionNotFound) {
-		return nil, fmt.Errorf("failed to fetch partition for response, error: %v", columnsResponse.GetErrorMessage())
-	}
-
-	return columnsResponse.GetColumns(), nil
-}
-
-func isPartitionedByEventDate(partitionedColumns []string) bool {
-	return slices.Contains(partitionedColumns, "event_date")
-}
-
-// partitionQuery
-// Checks whether the table is partition with event_date column
-// If specified, then calculates the date range from first and last event at and add it IN predicate query for event_date
-// If not specified, them returns empty string
-func (dl *Deltalake) partitionQuery(ctx context.Context, tableName string) (string, error) {
-	if !dl.EnablePartitionPruning {
-		return "", nil
-	}
-
-	partitionColumns, err := dl.fetchPartitionColumns(ctx, dl.Client, tableName)
-	if err != nil {
-		return "", fmt.Errorf("failed to prepare partition query, error: %w", err)
-	}
-
-	if !isPartitionedByEventDate(partitionColumns) {
-		return "", nil
-	}
-
-	firstEvent, lastEvent := dl.Uploader.GetFirstLastEvent()
-
-	dateRange := warehouseutils.GetDateRangeList(firstEvent, lastEvent, "2006-01-02")
-	if len(dateRange) == 0 {
-		return "", nil
-	}
-
-	dateRangeString := warehouseutils.JoinWithFormatting(dateRange, func(idx int, str string) string {
-		return fmt.Sprintf(`'%s'`, str)
-	}, ",")
-	query := fmt.Sprintf(`CAST ( MAIN.event_date AS string) IN (%s)`, dateRangeString)
-	return query, nil
-}
-
-// ExecuteSQLClient executes sql client using grpc Client
-func (*Deltalake) ExecuteSQLClient(ctx context.Context, client *client.Client, sqlStatement string) (err error) {
-	executeResponse, err := client.Client.Execute(ctx, &proto.ExecuteRequest{
-		Config:       client.CredConfig,
-		Identifier:   client.CredIdentifier,
-		SqlStatement: sqlStatement,
-	})
-	if err != nil {
-		return fmt.Errorf("error while executing: %v", err)
-	}
-	if !checkAndIgnoreAlreadyExistError(executeResponse.GetErrorCode(), databaseNotFound) || !checkAndIgnoreAlreadyExistError(executeResponse.GetErrorCode(), tableOrViewNotFound) {
-		err = fmt.Errorf("error while executing with response: %v", executeResponse.GetErrorMessage())
-		return
-	}
-	return
-}
-
-// schemaExists checks it schema exists or not.
-func (dl *Deltalake) schemaExists(ctx context.Context, schemaName string) (exists bool, err error) {
-	sqlStatement := fmt.Sprintf(`SHOW SCHEMAS LIKE '%s';`, schemaName)
-	fetchSchemasResponse, err := dl.Client.Client.FetchSchemas(ctx, &proto.FetchSchemasRequest{
-		Config:       dl.Client.CredConfig,
-		Identifier:   dl.Client.CredIdentifier,
-		SqlStatement: sqlStatement,
-	})
-	if err != nil {
-		return exists, fmt.Errorf("%s Error while fetching schemas: %v", dl.GetLogIdentifier(), err)
-	}
-	if !checkAndIgnoreAlreadyExistError(fetchSchemasResponse.GetErrorCode(), databaseNotFound) {
-		err = fmt.Errorf("%s Error while fetching schemas with response: %v", dl.GetLogIdentifier(), fetchSchemasResponse.GetErrorMessage())
-		return
-	}
-	exists = len(fetchSchemasResponse.GetDatabases()) == 1 && fetchSchemasResponse.GetDatabases()[0] == schemaName
-	return
-}
-
-// createSchema creates schema
-func (dl *Deltalake) createSchema(ctx context.Context) (err error) {
-	sqlStatement := fmt.Sprintf(`CREATE SCHEMA IF NOT EXISTS %s;`, dl.Namespace)
-	dl.Logger.Infof("%s Creating schema in delta lake with SQL:%v", dl.GetLogIdentifier(), sqlStatement)
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
-	return
-}
-
-// dropStagingTables drops staging tables
-func (dl *Deltalake) dropStagingTables(ctx context.Context, tableNames []string) {
-	for _, stagingTableName := range tableNames {
-		dl.Logger.Infof("%s Dropping table %+v\n", dl.GetLogIdentifier(), stagingTableName)
-		sqlStatement := fmt.Sprintf(`DROP TABLE %[1]s.%[2]s;`, dl.Namespace, stagingTableName)
-		dropTableResponse, err := dl.Client.Client.Execute(ctx, &proto.ExecuteRequest{
-			Config:       dl.Client.CredConfig,
-			Identifier:   dl.Client.CredIdentifier,
-			SqlStatement: sqlStatement,
-		})
-		if err != nil {
-			dl.Logger.Errorf("%s Error dropping staging tables in delta lake: %v", dl.GetLogIdentifier(), err)
-			continue
-		}
-		if !checkAndIgnoreAlreadyExistError(dropTableResponse.GetErrorCode(), tableOrViewNotFound) {
-			dl.Logger.Errorf("%s Error dropping staging tables in delta lake: %v", dl.GetLogIdentifier(), dropTableResponse.GetErrorMessage())
-		}
-	}
-}
-
-// sortedColumnNames returns sorted column names
-func (dl *Deltalake) sortedColumnNames(tableSchemaInUpload model.TableSchema, sortedColumnKeys []string, diff warehouseutils.TableSchemaDiff) (sortedColumnNames string) {
-	if dl.Uploader.GetLoadFileType() == warehouseutils.LOAD_FILE_TYPE_PARQUET {
-		sortedColumnNames = warehouseutils.JoinWithFormatting(sortedColumnKeys, func(_ int, value string) string {
-			columnName := value
-			columnType := dataTypesMap[tableSchemaInUpload[columnName]]
-			return fmt.Sprintf(`%s::%s`, columnName, columnType)
-		}, ",")
-	} else {
-		// TODO: Explore adding headers to csv.
-		format := func(index int, value string) string {
-			csvColumnIndex := fmt.Sprintf(`%s%d`, "_c", index)
-			columnName := value
-			columnType := getDeltaLakeDataType(tableSchemaInUpload[columnName])
-			return fmt.Sprintf(`CAST ( %s AS %s ) AS %s`, csvColumnIndex, columnType, columnName)
-		}
-		formatString := warehouseutils.JoinWithFormatting(sortedColumnKeys, format, ",")
-		if len(diff.ColumnMap) > 0 {
-			diffCols := make([]string, 0, len(diff.ColumnMap))
-			for key := range diff.ColumnMap {
-				diffCols = append(diffCols, key)
-			}
-			diffFormat := func(index int, value string) string {
-				return fmt.Sprintf(`NULL AS %s`, value)
-			}
-			diffString := warehouseutils.JoinWithFormatting(diffCols, diffFormat, ",")
-			return fmt.Sprintf("%s, %s", formatString, diffString)
-		}
-		return formatString
-	}
-	return
-}
-
-// credentialsStr return authentication for AWS STS and SSE-C encryption
-// STS authentication is only supported with S3A client.
-func (dl *Deltalake) credentialsStr() (string, error) {
-	if dl.ObjectStorage == warehouseutils.S3 {
-		canUseRudderStorage := misc.IsConfiguredToUseRudderObjectStorage(dl.Warehouse.Destination.Config)
-		canUseSTSTokens := warehouseutils.GetConfigValueBoolString(UseSTSTokens, dl.Warehouse) == "true"
-		if canUseRudderStorage || canUseSTSTokens {
-			tempAccessKeyId, tempSecretAccessKey, token, err := warehouseutils.GetTemporaryS3Cred(&dl.Warehouse.Destination)
-			if err != nil {
-				return "", fmt.Errorf("temporary s3 credentials: %w", err)
-			}
-			auth := fmt.Sprintf(`CREDENTIALS ( 'awsKeyId' = '%s', 'awsSecretKey' = '%s', 'awsSessionToken' = '%s' )`, tempAccessKeyId, tempSecretAccessKey, token)
-			return auth, nil
-		}
-	}
-	return "", nil
-}
-
-// getLoadFolder return the load folder where the load files are present
-func (dl *Deltalake) getLoadFolder(location string) (loadFolder string) {
-	loadFolder = warehouseutils.GetObjectFolderForDeltalake(dl.ObjectStorage, location)
-	if dl.ObjectStorage == warehouseutils.S3 {
-		awsAccessKey := warehouseutils.GetConfigValue(warehouseutils.AWSAccessKey, dl.Warehouse)
-		awsSecretKey := warehouseutils.GetConfigValue(warehouseutils.AWSAccessSecret, dl.Warehouse)
-		if awsAccessKey != "" && awsSecretKey != "" {
-			loadFolder = strings.Replace(loadFolder, "s3://", "s3a://", 1)
-		}
-	}
-	return
-}
-
-func getTableSchemaDiff(tableSchemaInUpload, tableSchemaAfterUpload model.TableSchema) (diff warehouseutils.TableSchemaDiff) {
-	diff = warehouseutils.TableSchemaDiff{
-		ColumnMap: make(model.TableSchema),
-	}
-	diff.ColumnMap = make(model.TableSchema)
-	for columnName, columnType := range tableSchemaAfterUpload {
-		if _, ok := tableSchemaInUpload[columnName]; !ok {
-			diff.ColumnMap[columnName] = columnType
-		}
-	}
-	return diff
-}
-
-// loadTable Loads table with table name
-func (dl *Deltalake) loadTable(ctx context.Context, tableName string, tableSchemaInUpload, tableSchemaAfterUpload model.TableSchema, skipTempTableDelete bool) (stagingTableName string, err error) {
-	// Getting sorted column keys from tableSchemaInUpload
-	sortedColumnKeys := warehouseutils.SortColumnKeysFromColumnMap(tableSchemaInUpload)
-
-	// Creating staging table
-	stagingTableName = warehouseutils.StagingTableName(provider, tableName, tableNameLimit)
-	err = dl.CreateTable(ctx, stagingTableName, tableSchemaAfterUpload)
-	if err != nil {
-		return
-	}
-
-	// Dropping staging tables if required
-	if !skipTempTableDelete {
-		defer dl.dropStagingTables(ctx, []string{stagingTableName})
-	}
-
-	// Get the credentials string to copy from the staging location to table
-	auth, err := dl.credentialsStr()
-	if err != nil {
-		return
-	}
-
-	// Getting objects location
-	objectsLocation, err := dl.Uploader.GetSampleLoadFileLocation(ctx, tableName)
-	if err != nil {
-		return
-	}
-
-	loadFolder := dl.getLoadFolder(objectsLocation)
-
-	// Creating copy sql statement to copy from load folder to the staging table
-	tableSchemaDiff := getTableSchemaDiff(tableSchemaInUpload, tableSchemaAfterUpload)
-	sortedColumnNames := dl.sortedColumnNames(tableSchemaInUpload, sortedColumnKeys, tableSchemaDiff)
-	var sqlStatement string
-	if dl.Uploader.GetLoadFileType() == warehouseutils.LOAD_FILE_TYPE_PARQUET {
-		sqlStatement = fmt.Sprintf("COPY INTO %v FROM ( SELECT %v FROM '%v' ) "+
-			"FILEFORMAT = PARQUET "+
-			"PATTERN = '*.parquet' "+
-			"COPY_OPTIONS ('force' = 'true') "+
-			"%s;",
-			fmt.Sprintf(`%s.%s`, dl.Namespace, stagingTableName), sortedColumnNames, loadFolder, auth)
-	} else {
-		sqlStatement = fmt.Sprintf("COPY INTO %v FROM ( SELECT %v FROM '%v' ) "+
-			"FILEFORMAT = CSV "+
-			"PATTERN = '*.gz' "+
-			"FORMAT_OPTIONS ( 'compression' = 'gzip', 'quote' = '\"', 'escape' = '\"', 'multiLine' = 'true' ) "+
-			"COPY_OPTIONS ('force' = 'true') "+
-			"%s;",
-			fmt.Sprintf(`%s.%s`, dl.Namespace, stagingTableName), sortedColumnNames, loadFolder, auth)
-	}
-
-	// Sanitising copy sql statement for logging
-	sanitisedSQLStmt, regexErr := misc.ReplaceMultiRegex(sqlStatement, map[string]string{
-		"'awsKeyId' = '[^']*'":        "'awsKeyId' = '***'",
-		"'awsSecretKey' = '[^']*'":    "'awsSecretKey' = '***'",
-		"'awsSessionToken' = '[^']*'": "'awsSessionToken' = '***'",
-	})
-	if regexErr == nil {
-		dl.Logger.Infof("%s Running COPY command with SQL: %s\n", dl.GetLogIdentifier(tableName), sanitisedSQLStmt)
-	}
-
-	// Executing copy sql statement
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
-	if err != nil {
-		dl.Logger.Errorf("%s Error running COPY command with SQL: %s\n error: %v", dl.GetLogIdentifier(tableName), sqlStatement, err)
-		return
-	}
-
-	if dl.LoadTableStrategy == "APPEND" {
-		sqlStatement = appendableLTSQLStatement(
-			dl.Namespace,
-			tableName,
-			stagingTableName,
-			warehouseutils.SortColumnKeysFromColumnMap(tableSchemaAfterUpload),
-		)
-	} else {
-		// Partition query
-		var partitionQuery string
-		partitionQuery, err = dl.partitionQuery(ctx, tableName)
-		if err != nil {
-			err = fmt.Errorf("failed getting partition query during load table, error: %w", err)
-			return
-		}
-
-		sqlStatement = mergeableLTSQLStatement(
-			dl.Namespace,
-			tableName,
-			stagingTableName,
-			sortedColumnKeys,
-			partitionQuery,
-		)
-	}
-	dl.Logger.Infof("%v Inserting records using staging table with SQL: %s\n", dl.GetLogIdentifier(tableName), sqlStatement)
-
-	// Executing load table sql statement
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
-	if err != nil {
-		dl.Logger.Errorf("%v Error inserting into original table: %v\n", dl.GetLogIdentifier(tableName), err)
-		return
-	}
-
-	dl.Logger.Infof("%v Complete load for table\n", dl.GetLogIdentifier(tableName))
-	return
-}
-
-// loadUserTables Loads users table
-func (dl *Deltalake) loadUserTables(ctx context.Context) (errorMap map[string]error) {
-	// Creating errorMap
-	errorMap = map[string]error{warehouseutils.IdentifiesTable: nil}
-	dl.Logger.Infof("%s Starting load for identifies and users tables\n", dl.GetLogIdentifier())
-
-	// Loading identifies tables
-	identifyStagingTable, err := dl.loadTable(ctx, warehouseutils.IdentifiesTable, dl.Uploader.GetTableSchemaInUpload(warehouseutils.IdentifiesTable), dl.Uploader.GetTableSchemaInWarehouse(warehouseutils.IdentifiesTable), true)
-	if err != nil {
-		errorMap[warehouseutils.IdentifiesTable] = err
-		return
-	}
-
-	// dropping identifies staging table
-	defer dl.dropStagingTables(ctx, []string{identifyStagingTable})
-
-	// Checking if users schema is present in GetTableSchemaInUpload
-	if len(dl.Uploader.GetTableSchemaInUpload(warehouseutils.UsersTable)) == 0 {
-		return
-	}
-	errorMap[warehouseutils.UsersTable] = nil
-
-	// Creating userColNames and firstValProps for create table sql statement
-	userColMap := dl.Uploader.GetTableSchemaInWarehouse(warehouseutils.UsersTable)
-	var userColNames, firstValProps []string
-	for colName := range userColMap {
-		// do not reference uuid in queries as it can be an auto incrementing field set by segment compatible tables
-		if colName == "id" || colName == "user_id" || colName == "uuid" {
-			continue
-		}
-		userColNames = append(userColNames, colName)
-		firstValProps = append(firstValProps, fmt.Sprintf(`FIRST_VALUE(%[1]s , TRUE) OVER (PARTITION BY id ORDER BY received_at DESC ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS %[1]s`, colName))
-	}
-	stagingTableName := warehouseutils.StagingTableName(provider, warehouseutils.UsersTable, tableNameLimit)
-
-	tableLocationSql := dl.getTableLocationSql(stagingTableName)
-
-	// Creating create table sql statement for staging users table
-	sqlStatement := fmt.Sprintf(`CREATE TABLE %[1]s.%[2]s USING DELTA %[7]s AS (SELECT DISTINCT * FROM
-										(
-											SELECT
-											id, %[3]s
-											FROM (
-												(
-													SELECT id, %[6]s FROM %[1]s.%[4]s WHERE id in (SELECT DISTINCT(user_id) FROM %[1]s.%[5]s WHERE user_id IS NOT NULL)
-												) UNION
-												(
-													SELECT user_id, %[6]s FROM %[1]s.%[5]s WHERE user_id IS NOT NULL
-												)
-											)
-										)
-									);`,
-		dl.Namespace,
-		stagingTableName,
-		strings.Join(firstValProps, ","),
-		warehouseutils.UsersTable,
-		identifyStagingTable,
-		columnNames(userColNames),
-		tableLocationSql,
+// Setup sets up the warehouse
+func (d *Deltalake) Setup(_ context.Context, warehouse model.Warehouse, uploader warehouseutils.Uploader) error {
+	d.Warehouse = warehouse
+	d.Namespace = warehouse.Namespace
+	d.Uploader = uploader
+	d.ObjectStorage = warehouseutils.ObjectStorageType(
+		warehouseutils.DELTALAKE,
+		warehouse.Destination.Config,
+		d.Uploader.UseRudderStorage(),
 	)
 
-	// Executing create sql statement
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
+	db, err := d.connect()
 	if err != nil {
-		dl.Logger.Errorf("%s Creating staging table for users failed with SQL: %s\n", dl.GetLogIdentifier(), sqlStatement)
-		dl.Logger.Errorf("%s Error creating users staging table from original table and identifies staging table: %v\n", dl.GetLogIdentifier(), err)
-		errorMap[warehouseutils.UsersTable] = err
+		return fmt.Errorf("connecting: %w", err)
+	}
+
+	d.DB = db
+
+	return nil
+}
+
+// connect connects to the warehouse
+func (d *Deltalake) connect() (*sqlmiddleware.DB, error) {
+	port, err := strconv.Atoi(warehouseutils.GetConfigValue(port, d.Warehouse))
+	if err != nil {
+		return nil, fmt.Errorf("port is not a number: %w", err)
+	}
+
+	connector, err := dbsql.NewConnector(
+		dbsql.WithServerHostname(warehouseutils.GetConfigValue(host, d.Warehouse)),
+		dbsql.WithPort(port),
+		dbsql.WithHTTPPath(warehouseutils.GetConfigValue(path, d.Warehouse)),
+		dbsql.WithAccessToken(warehouseutils.GetConfigValue(token, d.Warehouse)),
+		dbsql.WithSessionParams(map[string]string{
+			"ansi_mode": "false",
+		}),
+		dbsql.WithUserAgentEntry(userAgent),
+		dbsql.WithTimeout(d.ConnectTimeout),
+		dbsql.WithInitialNamespace(
+			warehouseutils.GetConfigValue(catalog, d.Warehouse),
+			"",
+		),
+		dbsql.WithRetries(d.MaxRetries, d.RetryMinWait, d.RetryMaxWait),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("creating connector: %w", err)
+	}
+
+	if err = dbsqllog.SetLogLevel("disabled"); err != nil {
+		return nil, fmt.Errorf("setting log level: %w", err)
+	}
+
+	db := sql.OpenDB(connector)
+	middleware := sqlmiddleware.New(
+		db,
+		sqlmiddleware.WithLogger(d.Logger),
+		sqlmiddleware.WithKeyAndValues(
+			logfield.SourceID, d.Warehouse.Source.ID,
+			logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+			logfield.DestinationID, d.Warehouse.Destination.ID,
+			logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+			logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+			logfield.Schema, d.Namespace,
+		),
+		sqlmiddleware.WithSlowQueryThreshold(d.SlowQueryThreshold),
+		sqlmiddleware.WithQueryTimeout(d.ConnectTimeout),
+		sqlmiddleware.WithSecretsRegex(map[string]string{
+			"'awsKeyId' = '[^']*'":        "'awsKeyId' = '***'",
+			"'awsSecretKey' = '[^']*'":    "'awsSecretKey' = '***'",
+			"'awsSessionToken' = '[^']*'": "'awsSessionToken' = '***'",
+		}),
+	)
+	return middleware, nil
+}
+
+// CrashRecover crash recover scenarios
+func (d *Deltalake) CrashRecover(ctx context.Context) {
+	d.dropDanglingStagingTables(ctx)
+}
+
+// dropDanglingStagingTables drops dangling staging tables
+func (d *Deltalake) dropDanglingStagingTables(ctx context.Context) {
+	tableNames, err := d.fetchTables(ctx, rudderStagingTableRegex)
+	if err != nil {
+		d.Logger.Warnw("fetching tables for dropping dangling staging tables",
+			logfield.SourceID, d.Warehouse.Source.ID,
+			logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+			logfield.DestinationID, d.Warehouse.Destination.ID,
+			logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+			logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+			logfield.Namespace, d.Namespace,
+			logfield.Error, err.Error(),
+		)
 		return
 	}
 
-	// Dropping staging users table
-	defer dl.dropStagingTables(ctx, []string{stagingTableName})
+	d.dropStagingTables(ctx, tableNames)
+}
 
-	// Creating the column Keys
-	columnKeys := append([]string{`id`}, userColNames...)
+// fetchTables fetches tables from the database
+func (d *Deltalake) fetchTables(ctx context.Context, regex string) ([]string, error) {
+	query := fmt.Sprintf(`SHOW tables FROM %s LIKE '%s';`, d.Namespace, regex)
 
-	if dl.LoadTableStrategy == "APPEND" {
-		sqlStatement = appendableLTSQLStatement(
-			dl.Namespace,
-			warehouseutils.UsersTable,
-			stagingTableName,
-			columnKeys,
+	rows, err := d.DB.QueryContext(ctx, query)
+	if err != nil {
+		if strings.Contains(err.Error(), schemaNotFound) {
+			return nil, nil
+		}
+		return nil, fmt.Errorf("executing fetching tables: %w", err)
+	}
+	defer func() { _ = rows.Close() }()
+
+	var tables []string
+	for rows.Next() {
+		var (
+			database    string
+			tableName   string
+			isTemporary bool
 		)
-	} else {
-		// Partition query
-		var partitionQuery string
-		partitionQuery, err = dl.partitionQuery(ctx, warehouseutils.UsersTable)
+
+		if err := rows.Scan(&database, &tableName, &isTemporary); err != nil {
+			return nil, fmt.Errorf("processing fetched tables: %w", err)
+		}
+
+		tables = append(tables, tableName)
+	}
+	if err := rows.Err(); err != nil {
+		return nil, fmt.Errorf("processing fetched tables: %w", err)
+	}
+	return tables, nil
+}
+
+// dropStagingTables drops all the staging tables
+func (d *Deltalake) dropStagingTables(ctx context.Context, stagingTables []string) {
+	for _, stagingTable := range stagingTables {
+		err := d.dropTable(ctx, stagingTable)
 		if err != nil {
-			err = fmt.Errorf("failed getting partition query during load users table, error: %w", err)
-			errorMap[warehouseutils.UsersTable] = err
-			return
+			d.Logger.Warnw("dropping staging table",
+				logfield.SourceID, d.Warehouse.Source.ID,
+				logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+				logfield.DestinationID, d.Warehouse.Destination.ID,
+				logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+				logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+				logfield.Namespace, d.Namespace,
+				logfield.StagingTableName, stagingTable,
+				logfield.Error, err.Error(),
+			)
 		}
-
-		sqlStatement = mergeableLTSQLStatement(
-			dl.Namespace,
-			warehouseutils.UsersTable,
-			stagingTableName,
-			columnKeys,
-			partitionQuery,
-		)
 	}
-	dl.Logger.Infof("%s Inserting records using staging table with SQL: %s\n", dl.GetLogIdentifier(warehouseutils.UsersTable), sqlStatement)
+}
 
-	// Executing the load users table sql statement
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
+// DropTable drops a table from the warehouse
+func (d *Deltalake) dropTable(ctx context.Context, table string) error {
+	query := fmt.Sprintf(`DROP TABLE %s.%s;`, d.Namespace, table)
+
+	_, err := d.DB.ExecContext(ctx, query)
 	if err != nil {
-		dl.Logger.Errorf("%s Error inserting into users table from staging table: %v\n", err)
-		errorMap[warehouseutils.UsersTable] = err
-		return
+		return fmt.Errorf("executing drop table: %w", err)
 	}
-	return
+
+	return nil
 }
 
-// getExternalLocation returns external location where we need to create the tables
-func (dl *Deltalake) getExternalLocation() (externalLocation string) {
-	enableExternalLocation := warehouseutils.GetConfigValueBoolString(EnableExternalLocation, dl.Warehouse)
-	if enableExternalLocation == "true" {
-		externalLocation := warehouseutils.GetConfigValue(ExternalLocation, dl.Warehouse)
-		return externalLocation
-	}
-	return
-}
-
-// getTableLocationSql returns external external table location
-func (dl *Deltalake) getTableLocationSql(tableName string) (tableLocation string) {
-	externalLocation := dl.getExternalLocation()
-	if externalLocation == "" {
-		return
-	}
-	return fmt.Sprintf("LOCATION '%s/%s/%s'", externalLocation, dl.Namespace, tableName)
-}
-
-// dropDanglingStagingTables drop dandling staging tables.
-func (dl *Deltalake) dropDanglingStagingTables(ctx context.Context) {
-	// Fetching the staging tables
-	tableNames, err := dl.fetchTables(ctx, dl.Client, dl.Namespace)
+// FetchSchema fetches the schema from the warehouse
+func (d *Deltalake) FetchSchema(ctx context.Context) (model.Schema, model.Schema, error) {
+	schema := make(model.Schema)
+	unrecognizedSchema := make(model.Schema)
+	tableNames, err := d.fetchTables(ctx, nonRudderStagingTableRegex)
 	if err != nil {
-		return
+		return model.Schema{}, model.Schema{}, fmt.Errorf("fetching tables: %w", err)
 	}
 
-	// Filtering tables based on not part of staging tables
-	var filteredTablesNames []string
+	// For each table, fetch the attributes
 	for _, tableName := range tableNames {
-		// Ignoring the staging tables
-		if !strings.HasPrefix(tableName, warehouseutils.StagingTablePrefix(provider)) {
-			continue
+		tableSchema, err := d.fetchTableAttributes(ctx, tableName)
+		if err != nil {
+			return model.Schema{}, model.Schema{}, fmt.Errorf("fetching table attributes: %w", err)
 		}
-		filteredTablesNames = append(filteredTablesNames, tableName)
-	}
 
-	// Drop staging tables
-	dl.dropStagingTables(ctx, filteredTablesNames)
+		for colName, dataType := range tableSchema {
+			if _, ok := excludeColumnsMap[colName]; ok {
+				continue
+			}
+
+			if _, ok := schema[tableName]; !ok {
+				schema[tableName] = make(model.TableSchema)
+			}
+			if datatype, ok := dataTypesMapToRudder[dataType]; ok {
+				schema[tableName][colName] = datatype
+			} else {
+				if _, ok := unrecognizedSchema[tableName]; !ok {
+					unrecognizedSchema[tableName] = make(model.TableSchema)
+				}
+				unrecognizedSchema[tableName][colName] = warehouseutils.MissingDatatype
+
+				warehouseutils.WHCounterStat(warehouseutils.RudderMissingDatatype, &d.Warehouse, warehouseutils.Tag{Name: "datatype", Value: datatype}).Count(1)
+			}
+		}
+	}
+	return schema, unrecognizedSchema, nil
 }
 
-// connectToWarehouse returns the database connection configured with Credentials
-func (dl *Deltalake) connectToWarehouse(ctx context.Context) (Client *client.Client, err error) {
-	credT := &client.Credentials{
-		Host:  warehouseutils.GetConfigValue(Host, dl.Warehouse),
-		Port:  warehouseutils.GetConfigValue(Port, dl.Warehouse),
-		Path:  warehouseutils.GetConfigValue(Path, dl.Warehouse),
-		Token: warehouseutils.GetConfigValue(Token, dl.Warehouse),
+// fetchTableAttributes fetches the attributes of a table
+func (d *Deltalake) fetchTableAttributes(ctx context.Context, tableName string) (model.TableSchema, error) {
+	tableSchema := make(model.TableSchema)
+
+	query := fmt.Sprintf(`DESCRIBE QUERY TABLE %s.%s;`, d.Namespace, tableName)
+
+	rows, err := d.DB.QueryContext(ctx, query)
+	if err != nil {
+		return nil, fmt.Errorf("executing fetching table attributes: %w", err)
 	}
-	return dl.NewClient(ctx, credT, dl.ConnectTimeout)
+	defer func() { _ = rows.Close() }()
+
+	for rows.Next() {
+		var (
+			colName, datatype string
+			comment           sql.NullString
+		)
+
+		if err = rows.Scan(&colName, &datatype, &comment); err != nil {
+			return nil, fmt.Errorf("processing fetched table attributes: %w", err)
+		}
+
+		tableSchema[colName] = datatype
+	}
+	if err = rows.Err(); err != nil {
+		return nil, fmt.Errorf("processing fetched table attributes: %w", err)
+	}
+	return tableSchema, nil
 }
 
-// CreateTable creates tables with table name and columns
-func (dl *Deltalake) CreateTable(ctx context.Context, tableName string, columns model.TableSchema) (err error) {
-	name := fmt.Sprintf(`%s.%s`, dl.Namespace, tableName)
+// CreateSchema creates a schema in the warehouse if it does not exist.
+func (d *Deltalake) CreateSchema(ctx context.Context) error {
+	if exists, err := d.schemaExists(ctx); err != nil {
+		return fmt.Errorf("checking if schema exists: %w", err)
+	} else if exists {
+		return nil
+	} else if err := d.createSchema(ctx); err != nil {
+		return fmt.Errorf("create schema: %w", err)
+	}
 
-	tableLocationSql := dl.getTableLocationSql(tableName)
-	var partitionedSql string
+	return nil
+}
+
+// schemaExists checks if a schema exists in the warehouse.
+func (d *Deltalake) schemaExists(ctx context.Context) (bool, error) {
+	query := fmt.Sprintf(`SHOW SCHEMAS LIKE '%s';`, d.Namespace)
+
+	var schema string
+	err := d.DB.QueryRowContext(ctx, query).Scan(&schema)
+
+	if err == sql.ErrNoRows {
+		return false, nil
+	} else if err != nil {
+		return false, fmt.Errorf("schema exists: %w", err)
+	}
+
+	return schema == d.Namespace, nil
+}
+
+// createSchema creates a schema in the warehouse.
+func (d *Deltalake) createSchema(ctx context.Context) error {
+	query := fmt.Sprintf(`CREATE SCHEMA IF NOT EXISTS %s;`, d.Namespace)
+
+	_, err := d.DB.ExecContext(ctx, query)
+	if err != nil {
+		return fmt.Errorf("executing create schema: %w", err)
+	}
+
+	return nil
+}
+
+// CreateTable creates a table in the warehouse.
+func (d *Deltalake) CreateTable(ctx context.Context, tableName string, columns model.TableSchema) error {
+	var partitionedSql, tableLocationSql string
+
+	tableLocationSql = d.tableLocationQuery(tableName)
 	if _, ok := columns["received_at"]; ok {
 		partitionedSql = `PARTITIONED BY(event_date)`
 	}
@@ -781,332 +444,365 @@ func (dl *Deltalake) CreateTable(ctx context.Context, tableName string, columns 
 		createTableClauseSql = "CREATE OR REPLACE TABLE"
 	}
 
-	sqlStatement := fmt.Sprintf(`%s %s ( %v ) USING DELTA %s %s;`, createTableClauseSql, name, ColumnsWithDataTypes(columns, ""), tableLocationSql, partitionedSql)
-	dl.Logger.Infof("%s Creating table in delta lake with SQL: %v", dl.GetLogIdentifier(tableName), sqlStatement)
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
-	return
-}
-
-func (dl *Deltalake) DropTable(ctx context.Context, tableName string) (err error) {
-	dl.Logger.Infof("%s Dropping table %s", dl.GetLogIdentifier(), tableName)
-	sqlStatement := fmt.Sprintf(`DROP TABLE %[1]s.%[2]s;`, dl.Namespace, tableName)
-	dropTableResponse, err := dl.Client.Client.Execute(ctx, &proto.ExecuteRequest{
-		Config:       dl.Client.CredConfig,
-		Identifier:   dl.Client.CredIdentifier,
-		SqlStatement: sqlStatement,
-	})
-	if err != nil {
-		return
-	}
-	if !checkAndIgnoreAlreadyExistError(dropTableResponse.GetErrorCode(), tableOrViewNotFound) {
-		err = fmt.Errorf("%s Error while dropping table with response: %v", dl.GetLogIdentifier(), dropTableResponse.GetErrorMessage())
-		return
-	}
-	return
-}
-
-func (dl *Deltalake) AddColumns(ctx context.Context, tableName string, columnsInfo []warehouseutils.ColumnInfo) error {
-	var (
-		query        string
-		queryBuilder strings.Builder
-		err          error
+	query := fmt.Sprintf(`
+		%s %s.%s ( %s ) USING DELTA %s %s;
+`,
+		createTableClauseSql,
+		d.Namespace,
+		tableName,
+		columnsWithDataTypes(columns, ""),
+		tableLocationSql,
+		partitionedSql,
 	)
+
+	_, err := d.DB.ExecContext(ctx, query)
+	if err != nil {
+		return fmt.Errorf("creating table: %w", err)
+	}
+
+	return nil
+}
+
+// columnsWithDataTypes returns the columns with their data types.
+func columnsWithDataTypes(columns model.TableSchema, prefix string) string {
+	keys := warehouseutils.SortColumnKeysFromColumnMap(columns)
+	format := func(_ int, name string) string {
+		if _, ok := excludeColumnsMap[name]; ok {
+			return ""
+		}
+		if name == "received_at" {
+			generatedColumnSQL := "DATE GENERATED ALWAYS AS ( CAST(received_at AS DATE) )"
+			return fmt.Sprintf(`%s%s %s, %s%s %s`, prefix, name, dataTypesMap[columns[name]], prefix, "event_date", generatedColumnSQL)
+		}
+		return fmt.Sprintf(`%s%s %s`, prefix, name, dataTypesMap[columns[name]])
+	}
+	return warehouseutils.JoinWithFormatting(keys, format, ",")
+}
+
+// tableLocationQuery returns the location query for the table.
+func (d *Deltalake) tableLocationQuery(tableName string) string {
+	enableExternalLocation := warehouseutils.GetConfigValueBoolString("enableExternalLocation", d.Warehouse)
+	externalLocation := warehouseutils.GetConfigValue("externalLocation", d.Warehouse)
+
+	if enableExternalLocation != "true" || externalLocation == "" {
+		return ""
+	}
+
+	return fmt.Sprintf("LOCATION '%s/%s/%s'", externalLocation, d.Namespace, tableName)
+}
+
+// AddColumns adds columns to the table.
+func (d *Deltalake) AddColumns(ctx context.Context, tableName string, columnsInfo []warehouseutils.ColumnInfo) error {
+	var queryBuilder strings.Builder
 
 	queryBuilder.WriteString(fmt.Sprintf(`
 		ALTER TABLE
 		  %s.%s
 		ADD COLUMNS(`,
-		dl.Namespace,
+		d.Namespace,
 		tableName,
 	))
 
 	for _, columnInfo := range columnsInfo {
-		queryBuilder.WriteString(fmt.Sprintf(` %s %s,`, columnInfo.Name, getDeltaLakeDataType(columnInfo.Type)))
+		queryBuilder.WriteString(fmt.Sprintf(` %s %s,`, columnInfo.Name, dataTypesMap[columnInfo.Type]))
 	}
 
-	query = strings.TrimSuffix(queryBuilder.String(), ",")
+	query := strings.TrimSuffix(queryBuilder.String(), ",")
 	query += ");"
 
-	dl.Logger.Infof("DL: Adding columns for destinationID: %s, tableName: %s with query: %v", dl.Warehouse.Destination.ID, tableName, query)
-	executeResponse, err := dl.Client.Client.Execute(ctx, &proto.ExecuteRequest{
-		Config:       dl.Client.CredConfig,
-		Identifier:   dl.Client.CredIdentifier,
-		SqlStatement: query,
-	})
-	if err != nil {
-		return fmt.Errorf("add columns: %w", err)
-	}
+	_, err := d.DB.ExecContext(ctx, query)
 
 	// Handle error in case of single column
 	if len(columnsInfo) == 1 {
-		if !checkAndIgnoreAlreadyExistError(executeResponse.GetErrorCode(), columnNotFound) {
-			return fmt.Errorf("add columns: %s", executeResponse.GetErrorMessage())
+		if err != nil && strings.Contains(err.Error(), columnsAlreadyExists) {
+			d.Logger.Infow("column already exists",
+				logfield.SourceID, d.Warehouse.Source.ID,
+				logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+				logfield.DestinationID, d.Warehouse.Destination.ID,
+				logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+				logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+				logfield.Namespace, d.Namespace,
+				logfield.TableName, tableName,
+				logfield.ColumnName, columnsInfo[0].Name,
+				logfield.Error, err.Error(),
+			)
+			return nil
 		}
-		return nil
 	}
 
-	if executeResponse.GetErrorCode() != "" {
-		return fmt.Errorf("add columns: %s", executeResponse.GetErrorMessage())
+	if err != nil {
+		return fmt.Errorf("adding columns: %w", err)
 	}
+
 	return nil
 }
 
-// CreateSchema checks if schema exists or not. If it does not exist, it creates the schema.
-func (dl *Deltalake) CreateSchema(ctx context.Context) (err error) {
-	// Checking if schema exists or not
-	var schemaExists bool
-	schemaExists, err = dl.schemaExists(ctx, dl.Namespace)
-	if err != nil {
-		dl.Logger.Errorf("%s Error checking if schema exists: %s, error: %v", dl.GetLogIdentifier(), dl.Namespace, err)
-		return err
-	}
-	if schemaExists {
-		dl.Logger.Infof("%s Skipping creating schema: %s since it already exists", dl.GetLogIdentifier(), dl.Namespace)
-		return
-	}
-
-	// Creating schema
-	return dl.createSchema(ctx)
-}
-
-// AlterColumn alter table with column name and type
+// AlterColumn alters a column in the warehouse
 func (*Deltalake) AlterColumn(context.Context, string, string, string) (model.AlterTableResponse, error) {
 	return model.AlterTableResponse{}, nil
 }
 
-// FetchSchema queries delta lake and returns the schema associated with provided namespace
-func (dl *Deltalake) FetchSchema(ctx context.Context) (model.Schema, model.Schema, error) {
-	// Schema Initialization
-	schema := make(model.Schema)
-	unrecognizedSchema := make(model.Schema)
+// LoadTable loads table for table name
+func (d *Deltalake) LoadTable(ctx context.Context, tableName string) error {
+	uploadTableSchema := d.Uploader.GetTableSchemaInUpload(tableName)
+	warehouseTableSchema := d.Uploader.GetTableSchemaInWarehouse(tableName)
 
-	// Fetching the tables
-	tableNames, err := dl.fetchTables(ctx, dl.Client, dl.Namespace)
+	_, err := d.loadTable(ctx, tableName, uploadTableSchema, warehouseTableSchema, false)
 	if err != nil {
-		return nil, nil, fmt.Errorf("fetching tables: %w", err)
+		return fmt.Errorf("loading table: %w", err)
 	}
 
-	// Filtering tables based on not part of staging tables
-	var filteredTablesNames []string
-	for _, tableName := range tableNames {
-		// Ignoring the staging tables
-		if strings.HasPrefix(tableName, warehouseutils.StagingTablePrefix(provider)) {
-			continue
-		}
-		filteredTablesNames = append(filteredTablesNames, tableName)
-	}
-
-	// For each table we are generating schema
-	for _, tableName := range filteredTablesNames {
-		fetchTableAttributesResponse, err := dl.Client.Client.FetchTableAttributes(ctx, &proto.FetchTableAttributesRequest{
-			Config:     dl.Client.CredConfig,
-			Identifier: dl.Client.CredIdentifier,
-			Schema:     dl.Namespace,
-			Table:      tableName,
-		})
-		if err != nil {
-			return schema, unrecognizedSchema, fmt.Errorf("%s Error while fetching table attributes: %v", dl.GetLogIdentifier(), err)
-		}
-		if !checkAndIgnoreAlreadyExistError(fetchTableAttributesResponse.GetErrorCode(), tableOrViewNotFound) {
-			return schema, unrecognizedSchema, fmt.Errorf("%s Error while fetching table attributes with response: %v", dl.GetLogIdentifier(), fetchTableAttributesResponse.GetErrorMessage())
-		}
-
-		// Populating the schema for the table
-		for _, item := range fetchTableAttributesResponse.GetAttributes() {
-			if _, ok := excludeColumnsMap[item.GetColName()]; ok {
-				continue
-			}
-
-			if _, ok := schema[tableName]; !ok {
-				schema[tableName] = make(model.TableSchema)
-			}
-			if datatype, ok := dataTypesMapToRudder[item.GetDataType()]; ok {
-				schema[tableName][item.GetColName()] = datatype
-			} else {
-				if _, ok := unrecognizedSchema[tableName]; !ok {
-					unrecognizedSchema[tableName] = make(model.TableSchema)
-				}
-				unrecognizedSchema[tableName][item.GetColName()] = warehouseutils.MISSING_DATATYPE
-
-				warehouseutils.WHCounterStat(warehouseutils.RUDDER_MISSING_DATATYPE, &dl.Warehouse, warehouseutils.Tag{Name: "datatype", Value: item.GetDataType()}).Count(1)
-			}
-		}
-	}
-	return schema, unrecognizedSchema, nil
-}
-
-// Setup populate the Deltalake
-func (dl *Deltalake) Setup(ctx context.Context, warehouse model.Warehouse, uploader warehouseutils.Uploader) (err error) {
-	dl.Warehouse = warehouse
-	dl.Namespace = warehouse.Namespace
-	dl.Uploader = uploader
-	dl.ObjectStorage = warehouseutils.ObjectStorageType(warehouseutils.DELTALAKE, warehouse.Destination.Config, dl.Uploader.UseRudderStorage())
-
-	dl.Client, err = dl.connectToWarehouse(ctx)
-	return err
-}
-
-// TestConnection test the connection for the warehouse
-func (*Deltalake) TestConnection(context.Context, model.Warehouse) error {
 	return nil
 }
 
-// Cleanup cleanup when upload is done.
-func (dl *Deltalake) Cleanup(ctx context.Context) {
-	if dl.Client != nil {
-		dl.dropDanglingStagingTables(ctx)
-		dl.Client.Close(ctx)
-	}
-}
-
-// CrashRecover crash recover scenarios
-func (dl *Deltalake) CrashRecover(ctx context.Context) {
-	dl.dropDanglingStagingTables(ctx)
-}
-
-// IsEmpty checks if the warehouse is empty or not
-func (*Deltalake) IsEmpty(context.Context, model.Warehouse) (empty bool, err error) {
-	return
-}
-
-// LoadUserTables loads user tables
-func (dl *Deltalake) LoadUserTables(ctx context.Context) map[string]error {
-	return dl.loadUserTables(ctx)
-}
-
-// LoadTable loads table for table name
-func (dl *Deltalake) LoadTable(ctx context.Context, tableName string) error {
-	_, err := dl.loadTable(ctx, tableName, dl.Uploader.GetTableSchemaInUpload(tableName), dl.Uploader.GetTableSchemaInWarehouse(tableName), false)
-	return err
-}
-
-// LoadIdentityMergeRulesTable loads identifies merge rules tables
-func (*Deltalake) LoadIdentityMergeRulesTable(context.Context) (err error) {
-	return
-}
-
-// LoadIdentityMappingsTable loads identifies mappings table
-func (*Deltalake) LoadIdentityMappingsTable(context.Context) (err error) {
-	return
-}
-
-// DownloadIdentityRules download identity rules
-func (*Deltalake) DownloadIdentityRules(context.Context, *misc.GZipWriter) (err error) {
-	return
-}
-
-// GetTotalCountInTable returns the total count in the table
-func (dl *Deltalake) GetTotalCountInTable(ctx context.Context, tableName string) (int64, error) {
+func (d *Deltalake) loadTable(ctx context.Context, tableName string, tableSchemaInUpload, tableSchemaAfterUpload model.TableSchema, skipTempTableDelete bool) (string, error) {
 	var (
-		err          error
-		sqlStatement string
+		sortedColumnKeys = warehouseutils.SortColumnKeysFromColumnMap(tableSchemaInUpload)
+		stagingTableName = warehouseutils.StagingTableName(provider, tableName, tableNameLimit)
+
+		err  error
+		auth string
+		row  *sqlmiddleware.Row
 	)
-	sqlStatement = fmt.Sprintf(`
-		SELECT COUNT(*) FROM %[1]s.%[2]s;
-	`,
-		dl.Namespace,
-		tableName,
+
+	d.Logger.Infow("started loading",
+		logfield.SourceID, d.Warehouse.Source.ID,
+		logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+		logfield.DestinationID, d.Warehouse.Destination.ID,
+		logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+		logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+		logfield.Namespace, d.Namespace,
+		logfield.TableName, tableName,
 	)
-	response, err := dl.Client.Client.FetchTotalCountInTable(ctx, &proto.FetchTotalCountInTableRequest{
-		Config:       dl.Client.CredConfig,
-		Identifier:   dl.Client.CredIdentifier,
-		SqlStatement: sqlStatement,
-	})
-	if err != nil {
-		return 0, fmt.Errorf("fetching table count: %w", err)
-	}
-	if response.GetErrorCode() != "" {
-		return 0, fmt.Errorf("fetching table count: %s", response.GetErrorMessage())
-	}
-	return response.GetCount(), nil
-}
 
-// Connect returns Client
-func (dl *Deltalake) Connect(ctx context.Context, warehouse model.Warehouse) (warehouseclient.Client, error) {
-	dl.Warehouse = warehouse
-	dl.Namespace = warehouse.Namespace
-	dl.ObjectStorage = warehouseutils.ObjectStorageType(
-		warehouseutils.DELTALAKE,
-		warehouse.Destination.Config,
-		misc.IsConfiguredToUseRudderObjectStorage(dl.Warehouse.Destination.Config),
+	if err = d.CreateTable(ctx, stagingTableName, tableSchemaAfterUpload); err != nil {
+		return "", fmt.Errorf("creating staging table: %w", err)
+	}
+
+	if !skipTempTableDelete {
+		defer d.dropStagingTables(ctx, []string{stagingTableName})
+	}
+
+	if auth, err = d.authQuery(); err != nil {
+		return "", fmt.Errorf("getting auth query: %w", err)
+	}
+
+	objectsLocation, err := d.Uploader.GetSampleLoadFileLocation(ctx, tableName)
+	if err != nil {
+		return "", fmt.Errorf("getting sample load file location: %w", err)
+	}
+
+	var (
+		loadFolder        = d.getLoadFolder(objectsLocation)
+		tableSchemaDiff   = tableSchemaDiff(tableSchemaInUpload, tableSchemaAfterUpload)
+		sortedColumnNames = d.sortedColumnNames(tableSchemaInUpload, sortedColumnKeys, tableSchemaDiff)
+
+		query          string
+		partitionQuery string
 	)
-	Client, err := dl.connectToWarehouse(ctx)
-	if err != nil {
-		return warehouseclient.Client{}, err
-	}
 
-	return warehouseclient.Client{Type: warehouseclient.DeltalakeClient, DeltalakeClient: Client}, err
-}
-
-// GetLogIdentifier returns log identifier
-func (dl *Deltalake) GetLogIdentifier(args ...string) string {
-	if len(args) == 0 {
-		return fmt.Sprintf("[%s][%s][%s][%s]", dl.Warehouse.Type, dl.Warehouse.Source.ID, dl.Warehouse.Destination.ID, dl.Warehouse.Namespace)
-	}
-	return fmt.Sprintf("[%s][%s][%s][%s][%s]", dl.Warehouse.Type, dl.Warehouse.Source.ID, dl.Warehouse.Destination.ID, dl.Warehouse.Namespace, strings.Join(args, "]["))
-}
-
-// GetDatabricksVersion Gets the databricks version by making a grpc call to Version stub.
-func GetDatabricksVersion(ctx context.Context) (databricksBuildVersion string) {
-	connectorURL := config.GetString("DATABRICKS_CONNECTOR_URL", "localhost:50051")
-
-	conn, err := grpc.DialContext(ctx, connectorURL, grpc.WithTransportCredentials(insecure.NewCredentials()))
-	if err != nil {
-		databricksBuildVersion = "Unable to create grpc connection to databricks."
-		return
-	}
-
-	versionClient := proto.NewVersionClient(conn)
-	versionResponse, err := versionClient.GetVersion(ctx, &proto.VersionRequest{})
-	if err != nil {
-		databricksBuildVersion = "Unable to read response from databricks."
-		return
-	}
-	databricksBuildVersion = versionResponse.GetVersion()
-	return
-}
-
-func (dl *Deltalake) LoadTestTable(ctx context.Context, location, tableName string, _ map[string]interface{}, format string) (err error) {
-	// Get the credentials string to copy from the staging location to table
-	auth, err := dl.credentialsStr()
-	if err != nil {
-		return
-	}
-
-	loadFolder := dl.getLoadFolder(location)
-
-	var sqlStatement string
-	if format == warehouseutils.LOAD_FILE_TYPE_PARQUET {
-		sqlStatement = fmt.Sprintf("COPY INTO %v FROM ( SELECT %v FROM '%v' ) "+
-			"FILEFORMAT = PARQUET "+
-			"PATTERN = '*.parquet' "+
-			"COPY_OPTIONS ('force' = 'true') "+
-			"%s;",
-			fmt.Sprintf(`%s.%s`, dl.Namespace, tableName),
-			fmt.Sprintf(`%s, %s`, "id", "val"),
-			loadFolder,
-			auth,
+	if d.Uploader.GetLoadFileType() == warehouseutils.LoadFileTypeParquet {
+		query = fmt.Sprintf(`
+			COPY INTO %s
+			FROM
+			  (
+				SELECT
+				  %s
+				FROM
+				  '%s'
+			  )
+			FILEFORMAT = PARQUET
+			PATTERN = '*.parquet'
+			COPY_OPTIONS ('force' = 'true')
+			%s;
+`,
+			fmt.Sprintf(`%s.%s`, d.Namespace, stagingTableName),
+			sortedColumnNames,
+			loadFolder, auth,
 		)
 	} else {
-		sqlStatement = fmt.Sprintf("COPY INTO %v FROM ( SELECT %v FROM '%v' ) "+
-			"FILEFORMAT = CSV "+
-			"PATTERN = '*.gz' "+
-			"FORMAT_OPTIONS ( 'compression' = 'gzip', 'quote' = '\"', 'escape' = '\"', 'multiLine' = 'true' ) "+
-			"COPY_OPTIONS ('force' = 'true') "+
-			"%s;",
-			fmt.Sprintf(`%s.%s`, dl.Namespace, tableName),
-			"CAST ( '_c0' AS BIGINT ) AS id, CAST ( '_c1' AS STRING ) AS val",
+		query = fmt.Sprintf(`
+			COPY INTO %s
+			FROM
+			  (
+				SELECT
+				  %s
+				FROM
+				  '%s'
+			  )
+			FILEFORMAT = CSV
+			PATTERN = '*.gz'
+			FORMAT_OPTIONS (
+				'compression' = 'gzip',
+				'quote' = '"',
+				'escape' = '"',
+				'multiLine' = 'true'
+			  )
+			COPY_OPTIONS ('force' = 'true')
+			%s;
+`,
+			fmt.Sprintf(`%s.%s`, d.Namespace, stagingTableName),
+			sortedColumnNames,
 			loadFolder,
 			auth,
 		)
 	}
 
-	err = dl.ExecuteSQLClient(ctx, dl.Client, sqlStatement)
-	return
+	if _, err = d.DB.ExecContext(ctx, query); err != nil {
+		return "", fmt.Errorf("running COPY command: %w", err)
+	}
+
+	if d.LoadTableStrategy == appendMode {
+		query = fmt.Sprintf(`
+			INSERT INTO %[1]s.%[2]s (%[4]s)
+			SELECT
+			  %[4]s
+			FROM
+			  (
+				SELECT
+				  *
+				FROM
+				  (
+					SELECT
+					  *,
+					  row_number() OVER (
+						PARTITION BY %[5]s
+						ORDER BY
+						  RECEIVED_AT DESC
+					  ) AS _rudder_staging_row_number
+					FROM
+					  %[1]s.%[3]s
+				  ) AS q
+				WHERE
+				  _rudder_staging_row_number = 1
+			  );
+		`,
+			d.Namespace,
+			tableName,
+			stagingTableName,
+			columnNames(warehouseutils.SortColumnKeysFromColumnMap(tableSchemaAfterUpload)),
+			primaryKey(tableName),
+		)
+	} else {
+		if partitionQuery, err = d.partitionQuery(ctx, tableName); err != nil {
+			return "", fmt.Errorf("getting partition query: %w", err)
+		}
+
+		pk := primaryKey(tableName)
+
+		query = fmt.Sprintf(`
+			MERGE INTO %[1]s.%[2]s AS MAIN USING (
+			  SELECT
+				*
+			  FROM
+				(
+				  SELECT
+					*,
+					row_number() OVER (
+					  PARTITION BY %[4]s
+					  ORDER BY
+						RECEIVED_AT DESC
+					) AS _rudder_staging_row_number
+				  FROM
+					%[1]s.%[3]s
+				) AS q
+			  WHERE
+				_rudder_staging_row_number = 1
+			)
+			AS STAGING ON %[8]s MAIN.%[4]s = STAGING.%[4]s
+			WHEN MATCHED THEN
+			UPDATE
+			SET
+			  %[5]s
+			WHEN NOT MATCHED THEN
+			INSERT (%[6]s)
+			VALUES
+			  (%[7]s);
+		`,
+			d.Namespace,
+			tableName,
+			stagingTableName,
+			pk,
+			columnsWithValues(sortedColumnKeys),
+			columnNames(sortedColumnKeys),
+			stagingColumnNames(sortedColumnKeys),
+			partitionQuery,
+		)
+	}
+
+	row = d.DB.QueryRowContext(ctx, query)
+
+	var (
+		affected int64
+		updated  int64
+		deleted  int64
+		inserted int64
+	)
+
+	if d.LoadTableStrategy == appendMode {
+		err = row.Scan(&affected, &inserted)
+	} else {
+		err = row.Scan(&affected, &updated, &deleted, &inserted)
+	}
+
+	if err != nil {
+		return "", fmt.Errorf("scanning deduplication: %w", err)
+	}
+	if row.Err() != nil {
+		return "", fmt.Errorf("running deduplication: %w", row.Err())
+	}
+
+	d.Stats.NewTaggedStat("dedup_rows", stats.CountType, stats.Tags{
+		"sourceID":    d.Warehouse.Source.ID,
+		"sourceType":  d.Warehouse.Source.SourceDefinition.Name,
+		"destID":      d.Warehouse.Destination.ID,
+		"destType":    d.Warehouse.Destination.DestinationDefinition.Name,
+		"workspaceId": d.Warehouse.WorkspaceID,
+		"tableName":   tableName,
+	}).Count(int(updated))
+
+	d.Logger.Infow("completed loading",
+		logfield.SourceID, d.Warehouse.Source.ID,
+		logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+		logfield.DestinationID, d.Warehouse.Destination.ID,
+		logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+		logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+		logfield.Namespace, d.Namespace,
+		logfield.TableName, tableName,
+	)
+	return stagingTableName, nil
 }
 
-func (dl *Deltalake) SetConnectionTimeout(timeout time.Duration) {
-	dl.ConnectTimeout = timeout
+func tableSchemaDiff(tableSchemaInUpload, tableSchemaAfterUpload model.TableSchema) warehouseutils.TableSchemaDiff {
+	diff := warehouseutils.TableSchemaDiff{
+		ColumnMap: make(model.TableSchema),
+	}
+
+	for columnName, columnType := range tableSchemaAfterUpload {
+		if _, ok := tableSchemaInUpload[columnName]; !ok {
+			diff.ColumnMap[columnName] = columnType
+		}
+	}
+
+	return diff
+}
+
+func columnNames(columns []string) string {
+	return strings.Join(columns, ",")
+}
+
+func stagingColumnNames(columns []string) string {
+	format := func(_ int, str string) string {
+		return fmt.Sprintf(`STAGING.%s`, str)
+	}
+	return warehouseutils.JoinWithFormatting(columns, format, ",")
+}
+
+func columnsWithValues(columns []string) string {
+	format := func(_ int, str string) string {
+		return fmt.Sprintf(`MAIN.%[1]s = STAGING.%[1]s`, str)
+	}
+	return warehouseutils.JoinWithFormatting(columns, format, ",")
 }
 
 func primaryKey(tableName string) string {
@@ -1117,94 +813,527 @@ func primaryKey(tableName string) string {
 	return key
 }
 
-func stagingSqlStatement(namespace, tableName, stagingTableName string, columnKeys []string) (sqlStatement string) {
-	pk := primaryKey(tableName)
-	if tableName == warehouseutils.UsersTable {
-		sqlStatement = fmt.Sprintf(`
+// sortedColumnNames returns the column names in the order of sortedColumnKeys
+func (d *Deltalake) sortedColumnNames(tableSchemaInUpload model.TableSchema, sortedColumnKeys []string, diff warehouseutils.TableSchemaDiff) string {
+	if d.Uploader.GetLoadFileType() == warehouseutils.LoadFileTypeParquet {
+		return warehouseutils.JoinWithFormatting(sortedColumnKeys, func(_ int, value string) string {
+			columnName := value
+			columnType := dataTypesMap[tableSchemaInUpload[columnName]]
+			return fmt.Sprintf(`%s::%s`, columnName, columnType)
+		}, ",")
+	}
+
+	format := func(index int, value string) string {
+		csvColumnIndex := fmt.Sprintf(`%s%d`, "_c", index)
+		columnName := value
+		columnType := dataTypesMap[tableSchemaInUpload[columnName]]
+		return fmt.Sprintf(`CAST ( %s AS %s ) AS %s`, csvColumnIndex, columnType, columnName)
+	}
+	formatString := warehouseutils.JoinWithFormatting(sortedColumnKeys, format, ",")
+
+	if len(diff.ColumnMap) > 0 {
+		diffCols := make([]string, 0, len(diff.ColumnMap))
+		for key := range diff.ColumnMap {
+			diffCols = append(diffCols, key)
+		}
+
+		diffFormat := func(_ int, value string) string {
+			return fmt.Sprintf(`NULL AS %s`, value)
+		}
+		diffString := warehouseutils.JoinWithFormatting(diffCols, diffFormat, ",")
+
+		return fmt.Sprintf("%s, %s", formatString, diffString)
+	}
+
+	return formatString
+}
+
+// authQuery return authentication for AWS STS and SSE-C encryption
+// STS authentication is only supported with S3A client.
+func (d *Deltalake) authQuery() (string, error) {
+	if d.ObjectStorage != warehouseutils.S3 || !d.canUseAuth() {
+		return "", nil
+	}
+
+	tempAccessKeyId, tempSecretAccessKey, token, err := warehouseutils.GetTemporaryS3Cred(&d.Warehouse.Destination)
+	if err != nil {
+		return "", fmt.Errorf("getting temporary s3 credentials: %w", err)
+	}
+
+	auth := fmt.Sprintf(`CREDENTIALS ( 'awsKeyId' = '%s', 'awsSecretKey' = '%s', 'awsSessionToken' = '%s' )`, tempAccessKeyId, tempSecretAccessKey, token)
+	return auth, nil
+}
+
+// canUseAuth returns true if the warehouse is configured to use RudderObjectStorage or STS tokens
+func (d *Deltalake) canUseAuth() bool {
+	canUseRudderStorage := misc.IsConfiguredToUseRudderObjectStorage(d.Warehouse.Destination.Config)
+	canUseSTSTokens := warehouseutils.GetConfigValueBoolString(useSTSTokens, d.Warehouse) == "true"
+
+	return canUseRudderStorage || canUseSTSTokens
+}
+
+// getLoadFolder returns the load folder for the warehouse load files
+func (d *Deltalake) getLoadFolder(location string) string {
+	loadFolder := warehouseutils.GetObjectFolderForDeltalake(d.ObjectStorage, location)
+
+	if d.ObjectStorage == warehouseutils.S3 && d.hasAWSCredentials() {
+		loadFolder = strings.Replace(loadFolder, "s3://", "s3a://", 1)
+	}
+
+	return loadFolder
+}
+
+// hasAWSCredentials returns true if the warehouse is configured to use AWS credentials
+func (d *Deltalake) hasAWSCredentials() bool {
+	awsAccessKey := warehouseutils.GetConfigValue(warehouseutils.AWSAccessKey, d.Warehouse)
+	awsSecretKey := warehouseutils.GetConfigValue(warehouseutils.AWSAccessSecret, d.Warehouse)
+
+	return awsAccessKey != "" && awsSecretKey != ""
+}
+
+// partitionQuery returns a query to fetch partitions for a table
+func (d *Deltalake) partitionQuery(ctx context.Context, tableName string) (string, error) {
+	if !d.EnablePartitionPruning {
+		return "", nil
+	}
+
+	query := fmt.Sprintf(`SHOW PARTITIONS %s.%s;`, d.Namespace, tableName)
+	rows, err := d.DB.QueryContext(ctx, query)
+	if err != nil {
+		if strings.Contains(err.Error(), partitionNotFound) {
+			return "", nil
+		}
+		return "", fmt.Errorf("executing fetching partitions: %w", err)
+	}
+	defer func() { _ = rows.Close() }()
+
+	_ = rows.Err() // ignore error
+
+	partitionColumns, err := rows.Columns()
+	if err != nil {
+		return "", fmt.Errorf("scanning partition columns: %w", err)
+	}
+
+	if !partitionedByEventDate(partitionColumns) {
+		return "", nil
+	}
+
+	firstEvent, lastEvent := d.Uploader.GetFirstLastEvent()
+	dateRange := warehouseutils.GetDateRangeList(firstEvent, lastEvent, "2006-01-02")
+	if len(dateRange) == 0 {
+		return "", nil
+	}
+
+	dateRangeString := warehouseutils.JoinWithFormatting(dateRange, func(_ int, str string) string {
+		return fmt.Sprintf(`'%s'`, str)
+	}, ",")
+	partitionQuery := fmt.Sprintf(`CAST ( MAIN.event_date AS string) IN (%s) AND`, dateRangeString)
+
+	return partitionQuery, nil
+}
+
+// partitionedByEventDate returns true if the table is partitioned by event_date
+func partitionedByEventDate(columns []string) bool {
+	return slices.Contains(columns, "event_date")
+}
+
+// LoadUserTables loads user tables
+func (d *Deltalake) LoadUserTables(ctx context.Context) map[string]error {
+	var (
+		identifiesSchemaInUpload    = d.Uploader.GetTableSchemaInUpload(warehouseutils.IdentifiesTable)
+		identifiesSchemaInWarehouse = d.Uploader.GetTableSchemaInWarehouse(warehouseutils.IdentifiesTable)
+		usersSchemaInUpload         = d.Uploader.GetTableSchemaInUpload(warehouseutils.UsersTable)
+		usersSchemaInWarehouse      = d.Uploader.GetTableSchemaInWarehouse(warehouseutils.UsersTable)
+	)
+
+	d.Logger.Infow("started loading for identifies and users tables",
+		logfield.SourceID, d.Warehouse.Source.ID,
+		logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+		logfield.DestinationID, d.Warehouse.Destination.ID,
+		logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+		logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+		logfield.Namespace, d.Namespace,
+	)
+
+	identifyStagingTable, err := d.loadTable(ctx, warehouseutils.IdentifiesTable, identifiesSchemaInUpload, identifiesSchemaInWarehouse, true)
+	if err != nil {
+		return map[string]error{
+			warehouseutils.IdentifiesTable: fmt.Errorf("loading table %s: %w", warehouseutils.IdentifiesTable, err),
+		}
+	}
+
+	defer d.dropStagingTables(ctx, []string{identifyStagingTable})
+
+	if len(usersSchemaInUpload) == 0 {
+		return map[string]error{
+			warehouseutils.IdentifiesTable: nil,
+		}
+	}
+
+	var (
+		partitionQuery string
+		row            *sqlmiddleware.Row
+	)
+
+	userColNames, firstValProps := getColumnProperties(usersSchemaInWarehouse)
+	stagingTableName := warehouseutils.StagingTableName(provider, warehouseutils.UsersTable, tableNameLimit)
+	tableLocationSql := d.tableLocationQuery(stagingTableName)
+
+	query := fmt.Sprintf(`
+		CREATE TABLE %[1]s.%[2]s USING DELTA %[7]s AS (
+		  SELECT
+			DISTINCT *
+		  FROM
+			(
+			  SELECT
+				id,
+				%[3]s
+			  FROM
+				(
+				  (
+					SELECT
+					  id,
+					  %[6]s
+					FROM
+					  %[1]s.%[4]s
+					WHERE
+					  id IN (
+						SELECT
+						  DISTINCT(user_id)
+						FROM
+						  %[1]s.%[5]s
+						WHERE
+						  user_id IS NOT NULL
+					  )
+				  )
+				  UNION
+					(
+					  SELECT
+						user_id,
+						%[6]s
+					  FROM
+						%[1]s.%[5]s
+					  WHERE
+						user_id IS NOT NULL
+					)
+				)
+			)
+		);
+`,
+		d.Namespace,
+		stagingTableName,
+		strings.Join(firstValProps, ","),
+		warehouseutils.UsersTable,
+		identifyStagingTable,
+		columnNames(userColNames),
+		tableLocationSql,
+	)
+
+	_, err = d.DB.ExecContext(ctx, query)
+
+	if err != nil {
+		return map[string]error{
+			warehouseutils.IdentifiesTable: nil,
+			warehouseutils.UsersTable:      fmt.Errorf("creating staging table for users: %w", err),
+		}
+	}
+
+	defer d.dropStagingTables(ctx, []string{stagingTableName})
+
+	columnKeys := append([]string{`id`}, userColNames...)
+
+	if d.LoadTableStrategy == appendMode {
+		query = fmt.Sprintf(`
+			INSERT INTO %[1]s.%[2]s (%[4]s)
 			SELECT
-			  %[3]s
+			  %[4]s
 			FROM
-			  %[1]s.%[2]s
+			  (
+				SELECT
+				  %[4]s
+				FROM
+				  %[1]s.%[3]s
+			  );
 		`,
-			namespace,
+			d.Namespace,
+			warehouseutils.UsersTable,
 			stagingTableName,
 			columnNames(columnKeys),
 		)
 	} else {
-		sqlStatement = fmt.Sprintf(`
-			SELECT
-			  *
+		if partitionQuery, err = d.partitionQuery(ctx, warehouseutils.UsersTable); err != nil {
+			return map[string]error{
+				warehouseutils.IdentifiesTable: nil,
+				warehouseutils.UsersTable:      fmt.Errorf("getting partition query: %w", err),
+			}
+		}
+
+		pk := primaryKey(warehouseutils.UsersTable)
+
+		query = fmt.Sprintf(`
+			MERGE INTO %[1]s.%[2]s AS MAIN USING (
+			  SELECT
+				%[6]s
+			  FROM
+				%[1]s.%[3]s
+			) AS STAGING ON %[8]s MAIN.%[4]s = STAGING.%[4]s
+			WHEN MATCHED THEN
+			UPDATE
+			SET
+			  %[5]s WHEN NOT MATCHED
+			THEN INSERT (%[6]s)
+			VALUES
+			  (%[7]s);
+		`,
+			d.Namespace,
+			warehouseutils.UsersTable,
+			stagingTableName,
+			pk,
+			columnsWithValues(columnKeys),
+			columnNames(columnKeys),
+			stagingColumnNames(columnKeys),
+			partitionQuery,
+		)
+	}
+
+	row = d.DB.QueryRowContext(ctx, query)
+
+	var (
+		affected int64
+		updated  int64
+		deleted  int64
+		inserted int64
+	)
+
+	if d.LoadTableStrategy == appendMode {
+		err = row.Scan(&affected, &inserted)
+	} else {
+		err = row.Scan(&affected, &updated, &deleted, &inserted)
+	}
+
+	if err != nil {
+		return map[string]error{
+			warehouseutils.IdentifiesTable: nil,
+			warehouseutils.UsersTable:      fmt.Errorf("getting rows affected for dedup: %w", err),
+		}
+	}
+
+	if row.Err() != nil {
+		return map[string]error{
+			warehouseutils.IdentifiesTable: nil,
+			warehouseutils.UsersTable:      fmt.Errorf("running deduplication: %w", row.Err()),
+		}
+	}
+
+	d.Stats.NewTaggedStat("dedup_rows", stats.CountType, stats.Tags{
+		"sourceID":    d.Warehouse.Source.ID,
+		"sourceType":  d.Warehouse.Source.SourceDefinition.Name,
+		"destID":      d.Warehouse.Destination.ID,
+		"destType":    d.Warehouse.Destination.DestinationDefinition.Name,
+		"workspaceId": d.Warehouse.WorkspaceID,
+		"tableName":   warehouseutils.UsersTable,
+	}).Count(int(updated))
+
+	d.Logger.Infow("completed loading for users and identifies tables",
+		logfield.SourceID, d.Warehouse.Source.ID,
+		logfield.SourceType, d.Warehouse.Source.SourceDefinition.Name,
+		logfield.DestinationID, d.Warehouse.Destination.ID,
+		logfield.DestinationType, d.Warehouse.Destination.DestinationDefinition.Name,
+		logfield.WorkspaceID, d.Warehouse.WorkspaceID,
+		logfield.Namespace, d.Namespace,
+	)
+
+	return map[string]error{
+		warehouseutils.IdentifiesTable: nil,
+		warehouseutils.UsersTable:      nil,
+	}
+}
+
+// getColumnProperties returns the column names and first value properties for the given table schema
+func getColumnProperties(usersSchemaInWarehouse model.TableSchema) ([]string, []string) {
+	var (
+		userColNames    []string
+		firstValProps   []string
+		ignoredColNames = map[string]struct{}{
+			"id":      {},
+			"user_id": {},
+			"uuid":    {},
+		}
+	)
+
+	for colName := range usersSchemaInWarehouse {
+		if _, ignore := ignoredColNames[colName]; ignore {
+			continue
+		}
+
+		userColNames = append(userColNames, colName)
+		firstValProps = append(firstValProps, fmt.Sprintf(`FIRST_VALUE(%[1]s, TRUE) OVER (PARTITION BY id ORDER BY received_at DESC ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS %[1]s`, colName))
+	}
+
+	return userColNames, firstValProps
+}
+
+// LoadIdentityMergeRulesTable loads identifies merge rules tables
+func (*Deltalake) LoadIdentityMergeRulesTable(context.Context) error {
+	return nil
+}
+
+// LoadIdentityMappingsTable loads identifies mappings table
+func (*Deltalake) LoadIdentityMappingsTable(context.Context) error {
+	return nil
+}
+
+// Cleanup cleans up the warehouse
+func (d *Deltalake) Cleanup(ctx context.Context) {
+	if d.DB != nil {
+		d.dropDanglingStagingTables(ctx)
+		_ = d.DB.Close()
+	}
+}
+
+// IsEmpty checks if the warehouse is empty or not
+func (*Deltalake) IsEmpty(context.Context, model.Warehouse) (bool, error) {
+	return false, nil
+}
+
+// TestConnection tests the connection to the warehouse
+func (d *Deltalake) TestConnection(ctx context.Context, _ model.Warehouse) error {
+	err := d.DB.PingContext(ctx)
+	if errors.Is(err, context.DeadlineExceeded) {
+		return fmt.Errorf("connection timeout: %w", err)
+	}
+	if err != nil {
+		return fmt.Errorf("pinging: %w", err)
+	}
+
+	return nil
+}
+
+// DownloadIdentityRules downloadchecking if schema exists identity rules
+func (*Deltalake) DownloadIdentityRules(context.Context, *misc.GZipWriter) error {
+	return nil
+}
+
+// GetTotalCountInTable returns the total count in the table
+func (d *Deltalake) GetTotalCountInTable(ctx context.Context, tableName string) (int64, error) {
+	query := fmt.Sprintf(`
+		SELECT COUNT(*) FROM %[1]s.%[2]s;
+	`,
+		d.Namespace,
+		tableName,
+	)
+
+	var total int64
+	err := d.DB.QueryRowContext(ctx, query).Scan(&total)
+	if err != nil {
+		if strings.Contains(err.Error(), schemaNotFound) {
+			return 0, nil
+		}
+		return 0, fmt.Errorf("total count in table: %w", err)
+	}
+
+	return total, nil
+}
+
+// Connect returns Client
+func (d *Deltalake) Connect(_ context.Context, warehouse model.Warehouse) (warehouseclient.Client, error) {
+	d.Warehouse = warehouse
+	d.Namespace = warehouse.Namespace
+	d.ObjectStorage = warehouseutils.ObjectStorageType(
+		warehouseutils.DELTALAKE,
+		warehouse.Destination.Config,
+		misc.IsConfiguredToUseRudderObjectStorage(d.Warehouse.Destination.Config),
+	)
+
+	db, err := d.connect()
+	if err != nil {
+		return warehouseclient.Client{}, fmt.Errorf("connecting: %w", err)
+	}
+
+	return warehouseclient.Client{Type: warehouseclient.SQLClient, SQL: db.DB}, nil
+}
+
+// LoadTestTable loads the test table
+func (d *Deltalake) LoadTestTable(ctx context.Context, location, tableName string, _ map[string]interface{}, format string) error {
+	auth, err := d.authQuery()
+	if err != nil {
+		return fmt.Errorf("auth query: %w", err)
+	}
+
+	loadFolder := d.getLoadFolder(location)
+
+	var query string
+	if format == warehouseutils.LoadFileTypeParquet {
+		query = fmt.Sprintf(`
+			COPY INTO %s
 			FROM
 			  (
 				SELECT
-				  *,
-				  row_number() OVER (
-					PARTITION BY %[3]s
-					ORDER BY
-					  RECEIVED_AT DESC
-				  ) AS _rudder_staging_row_number
+				  %s
 				FROM
-				  %[1]s.%[2]s
-			  ) AS q
-			WHERE
-			  _rudder_staging_row_number = 1
-		`,
-			namespace,
-			stagingTableName,
-			pk,
+				  '%s'
+			  )
+			FILEFORMAT = PARQUET
+			PATTERN = '*.parquet'
+			COPY_OPTIONS ('force' = 'true')
+			%s;
+`,
+			fmt.Sprintf(`%s.%s`, d.Namespace, tableName),
+			fmt.Sprintf(`%s, %s`, "id", "val"),
+			loadFolder,
+			auth,
+		)
+	} else {
+		query = fmt.Sprintf(`
+			COPY INTO %s
+			FROM
+			  (
+				SELECT
+				  %s
+				FROM
+				  '%s'
+			  )
+			FILEFORMAT = CSV
+			PATTERN = '*.gz'
+			FORMAT_OPTIONS (
+				'compression' = 'gzip',
+				'quote' = '"',
+				'escape' = '"',
+				'multiLine' = 'true'
+			)
+			COPY_OPTIONS ('force' = 'true')
+			%s;
+`,
+			fmt.Sprintf(`%s.%s`, d.Namespace, tableName),
+			"CAST ( '_c0' AS BIGINT ) AS id, CAST ( '_c1' AS STRING ) AS val",
+			loadFolder,
+			auth,
 		)
 	}
-	return
-}
 
-func mergeableLTSQLStatement(namespace, tableName, stagingTableName string, columnKeys []string, partitionQuery string) string {
-	pk := primaryKey(tableName)
-	if partitionQuery != "" {
-		partitionQuery += " AND"
+	_, err = d.DB.ExecContext(ctx, query)
+	if err != nil {
+		return fmt.Errorf("loading test table: %w", err)
 	}
-	stagingTableSqlStatement := stagingSqlStatement(namespace, tableName, stagingTableName, columnKeys)
-	sqlStatement := fmt.Sprintf(`
-		MERGE INTO %[1]s.%[2]s AS MAIN USING (%[3]s) AS STAGING ON %[8]s MAIN.%[4]s = STAGING.%[4]s
-		WHEN MATCHED THEN
-		UPDATE
-		SET
-		  %[5]s
-		  WHEN NOT MATCHED THEN
-		INSERT
-		  (%[6]s)
-		VALUES
-		  (%[7]s);
-		`,
-		namespace,
-		tableName,
-		stagingTableSqlStatement,
-		pk,
-		columnsWithValues(columnKeys),
-		columnNames(columnKeys),
-		stagingColumnNames(columnKeys),
-		partitionQuery,
-	)
-	return sqlStatement
+
+	return nil
 }
 
-func appendableLTSQLStatement(namespace, tableName, stagingTableName string, columnKeys []string) string {
-	stagingTableSqlStatement := stagingSqlStatement(namespace, tableName, stagingTableName, columnKeys)
-	sqlStatement := fmt.Sprintf(`
-		INSERT INTO %[1]s.%[2]s (%[4]s)
-		SELECT
-		  %[4]s
-		FROM
-		  (%[5]s);
-		`,
-		namespace,
-		tableName,
-		stagingTableName,
-		columnNames(columnKeys),
-		stagingTableSqlStatement,
-	)
-	return sqlStatement
+// SetConnectionTimeout sets the connection timeout
+func (d *Deltalake) SetConnectionTimeout(timeout time.Duration) {
+	d.ConnectTimeout = timeout
 }
 
+// ErrorMappings returns the error mappings
 func (*Deltalake) ErrorMappings() []model.JobError {
 	return errorsMappings
+}
+
+// DropTable drops a table in the warehouse
+func (d *Deltalake) DropTable(ctx context.Context, tableName string) error {
+	return d.dropTable(ctx, tableName)
+}
+
+func (*Deltalake) DeleteBy(context.Context, []string, warehouseutils.DeleteByParams) error {
+	return fmt.Errorf(warehouseutils.NotImplementedErrorCode)
 }
