+++ warehouse/integrations/deltalake/testdata/template.json
