+++ warehouse/integrations/deltalake/testdata/upload-job-append-mode.staging-1.json
