+++ warehouse/integrations/deltalake/testdata/upload-job-append-mode.staging-2.json
