+++ warehouse/integrations/deltalake/testdata/upload-job-parquet.staging-2.json
