+++ warehouse/internal/loadfiles/loadfiles.go
@@ -364,7 +364,7 @@ func (lf *LoadFileGenerator) destinationRevisionIDMap(ctx context.Context, job *
 
 func GetLoadFilePrefix(timeWindow time.Time, warehouse model.Warehouse) string {
 	switch warehouse.Type {
-	case warehouseutils.GCS_DATALAKE:
+	case warehouseutils.GCSDatalake:
 		windowFormat := timeWindow.Format(warehouseutils.DatalakeTimeWindowFormat)
 
 		if windowLayout := warehouseutils.GetConfigValue("timeWindowLayout", warehouse); windowLayout != "" {
@@ -374,7 +374,7 @@ func GetLoadFilePrefix(timeWindow time.Time, warehouse model.Warehouse) string {
 			windowFormat = fmt.Sprintf("%v/%v", suffix, windowFormat)
 		}
 		return windowFormat
-	case warehouseutils.S3_DATALAKE:
+	case warehouseutils.S3Datalake:
 		if !schemarepository.UseGlue(&warehouse) {
 			return timeWindow.Format(warehouseutils.DatalakeTimeWindowFormat)
 		}
