+++ warehouse/internal/loadfiles/loadfiles_test.go
@@ -372,7 +372,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 				Destination: backendconfig.DestinationT{
 					Config: map[string]interface{}{},
 				},
-				Type: warehouseutils.S3_DATALAKE,
+				Type: warehouseutils.S3Datalake,
 			},
 			expected: "2022/08/06/14",
 		},
@@ -385,7 +385,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"useGlue": true,
 					},
 				},
-				Type: warehouseutils.S3_DATALAKE,
+				Type: warehouseutils.S3Datalake,
 			},
 			expected: "2022/08/06/14",
 		},
@@ -399,7 +399,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"timeWindowLayout": "dt=2006-01-02",
 					},
 				},
-				Type: warehouseutils.S3_DATALAKE,
+				Type: warehouseutils.S3Datalake,
 			},
 			expected: "dt=2022-08-06",
 		},
@@ -411,7 +411,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"tableSuffix": "key=val",
 					},
 				},
-				Type: warehouseutils.AZURE_DATALAKE,
+				Type: warehouseutils.AzureDatalake,
 			},
 			expected: "2022/08/06/14",
 		},
@@ -421,7 +421,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 				Destination: backendconfig.DestinationT{
 					Config: map[string]interface{}{},
 				},
-				Type: warehouseutils.GCS_DATALAKE,
+				Type: warehouseutils.GCSDatalake,
 			},
 			expected: "2022/08/06/14",
 		},
@@ -433,7 +433,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"tableSuffix": "key=val",
 					},
 				},
-				Type: warehouseutils.GCS_DATALAKE,
+				Type: warehouseutils.GCSDatalake,
 			},
 			expected: "key=val/2022/08/06/14",
 		},
@@ -445,7 +445,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"timeWindowLayout": "year=2006/month=01/day=02/hour=15",
 					},
 				},
-				Type: warehouseutils.GCS_DATALAKE,
+				Type: warehouseutils.GCSDatalake,
 			},
 			expected: "year=2022/month=08/day=06/hour=14",
 		},
@@ -458,7 +458,7 @@ func TestGetLoadFilePrefix(t *testing.T) {
 						"timeWindowLayout": "year=2006/month=01/day=02/hour=15",
 					},
 				},
-				Type: warehouseutils.GCS_DATALAKE,
+				Type: warehouseutils.GCSDatalake,
 			},
 			expected: "key=val/year=2022/month=08/day=06/hour=14",
 		},
