+++ warehouse/logfield/logfield.go
@@ -23,10 +23,7 @@ const (
 	LoadFileType               = "loadFileType"
 	ErrorMapping               = "errorMapping"
 	DestinationCredsValid      = "destinationCredsValid"
-	Provider                   = "provider"
 	Query                      = "query"
 	QueryExecutionTime         = "queryExecutionTime"
 	StagingTableName           = "stagingTableName"
-	QueryPlanner               = "queryPlan"
-	LoadFile                   = "loadFile"
 )
