+++ warehouse/slave.go
@@ -330,7 +330,7 @@ func (jobRun *JobRun) GetWriter(tableName string) (encoding.LoadFileWriter, erro
 	if !ok {
 		var err error
 		outputFilePath := jobRun.getLoadFilePath(tableName)
-		if jobRun.job.LoadFileType == warehouseutils.LOAD_FILE_TYPE_PARQUET {
+		if jobRun.job.LoadFileType == warehouseutils.LoadFileTypeParquet {
 			writer, err = encoding.CreateParquetWriter(jobRun.job.UploadSchema[tableName], outputFilePath, jobRun.job.DestinationType)
 		} else {
 			writer, err = misc.CreateGZ(outputFilePath)
@@ -458,7 +458,7 @@ func processStagingFile(ctx context.Context, job Payload, workerIndex int) (load
 		tableName := batchRouterEvent.Metadata.Table
 		columnData := batchRouterEvent.Data
 
-		if job.DestinationType == warehouseutils.S3_DATALAKE && len(sortedTableColumnMap[tableName]) > columnCountLimitMap[warehouseutils.S3_DATALAKE] {
+		if job.DestinationType == warehouseutils.S3Datalake && len(sortedTableColumnMap[tableName]) > columnCountLimitMap[warehouseutils.S3Datalake] {
 			pkgLogger.Errorf("[WH]: Huge staging file columns : columns in upload schema: %v for StagingFileID: %v", len(sortedTableColumnMap[tableName]), job.StagingFileID)
 			return nil, fmt.Errorf("staging file schema limit exceeded for stagingFileID: %d, actualCount: %d", job.StagingFileID, len(sortedTableColumnMap[tableName]))
 		}
