+++ warehouse/upload.go
@@ -151,25 +151,25 @@ func init() {
 
 func setMaxParallelLoads() {
 	maxParallelLoads = map[string]int{
-		warehouseutils.BQ:             config.GetInt("Warehouse.bigquery.maxParallelLoads", 20),
-		warehouseutils.RS:             config.GetInt("Warehouse.redshift.maxParallelLoads", 8),
-		warehouseutils.POSTGRES:       config.GetInt("Warehouse.postgres.maxParallelLoads", 8),
-		warehouseutils.MSSQL:          config.GetInt("Warehouse.mssql.maxParallelLoads", 8),
-		warehouseutils.SNOWFLAKE:      config.GetInt("Warehouse.snowflake.maxParallelLoads", 8),
-		warehouseutils.CLICKHOUSE:     config.GetInt("Warehouse.clickhouse.maxParallelLoads", 8),
-		warehouseutils.DELTALAKE:      config.GetInt("Warehouse.deltalake.maxParallelLoads", 8),
-		warehouseutils.S3_DATALAKE:    config.GetInt("Warehouse.s3_datalake.maxParallelLoads", 8),
-		warehouseutils.GCS_DATALAKE:   config.GetInt("Warehouse.gcs_datalake.maxParallelLoads", 8),
-		warehouseutils.AZURE_DATALAKE: config.GetInt("Warehouse.azure_datalake.maxParallelLoads", 8),
+		warehouseutils.BQ:            config.GetInt("Warehouse.bigquery.maxParallelLoads", 20),
+		warehouseutils.RS:            config.GetInt("Warehouse.redshift.maxParallelLoads", 8),
+		warehouseutils.POSTGRES:      config.GetInt("Warehouse.postgres.maxParallelLoads", 8),
+		warehouseutils.MSSQL:         config.GetInt("Warehouse.mssql.maxParallelLoads", 8),
+		warehouseutils.SNOWFLAKE:     config.GetInt("Warehouse.snowflake.maxParallelLoads", 8),
+		warehouseutils.CLICKHOUSE:    config.GetInt("Warehouse.clickhouse.maxParallelLoads", 8),
+		warehouseutils.DELTALAKE:     config.GetInt("Warehouse.deltalake.maxParallelLoads", 8),
+		warehouseutils.S3Datalake:    config.GetInt("Warehouse.s3_datalake.maxParallelLoads", 8),
+		warehouseutils.GCSDatalake:   config.GetInt("Warehouse.gcs_datalake.maxParallelLoads", 8),
+		warehouseutils.AzureDatalake: config.GetInt("Warehouse.azure_datalake.maxParallelLoads", 8),
 	}
 	columnCountLimitMap = map[string]int{
-		warehouseutils.AZURE_SYNAPSE: config.GetInt("Warehouse.azure_synapse.columnCountLimit", 1024),
-		warehouseutils.BQ:            config.GetInt("Warehouse.bigquery.columnCountLimit", 10000),
-		warehouseutils.CLICKHOUSE:    config.GetInt("Warehouse.clickhouse.columnCountLimit", 1000),
-		warehouseutils.MSSQL:         config.GetInt("Warehouse.mssql.columnCountLimit", 1024),
-		warehouseutils.POSTGRES:      config.GetInt("Warehouse.postgres.columnCountLimit", 1600),
-		warehouseutils.RS:            config.GetInt("Warehouse.redshift.columnCountLimit", 1600),
-		warehouseutils.S3_DATALAKE:   config.GetInt("Warehouse.s3_datalake.columnCountLimit", 10000),
+		warehouseutils.AzureSynapse: config.GetInt("Warehouse.azure_synapse.columnCountLimit", 1024),
+		warehouseutils.BQ:           config.GetInt("Warehouse.bigquery.columnCountLimit", 10000),
+		warehouseutils.CLICKHOUSE:   config.GetInt("Warehouse.clickhouse.columnCountLimit", 1000),
+		warehouseutils.MSSQL:        config.GetInt("Warehouse.mssql.columnCountLimit", 1024),
+		warehouseutils.POSTGRES:     config.GetInt("Warehouse.postgres.columnCountLimit", 1600),
+		warehouseutils.RS:           config.GetInt("Warehouse.redshift.columnCountLimit", 1600),
+		warehouseutils.S3Datalake:   config.GetInt("Warehouse.s3_datalake.columnCountLimit", 10000),
 	}
 }
 
@@ -1136,7 +1136,7 @@ func (job *UploadJob) columnCountStat(tableName string) {
 	)
 
 	switch job.warehouse.Type {
-	case warehouseutils.S3_DATALAKE, warehouseutils.GCS_DATALAKE, warehouseutils.AZURE_DATALAKE:
+	case warehouseutils.S3Datalake, warehouseutils.GCSDatalake, warehouseutils.AzureDatalake:
 		return
 	}
